Private API

Mesh

LFAToolkit.Mesh — Type

Rectangular mesh with independent scaling in each dimesion

Basis

LFAToolkit.Basis — Type

Finite element basis for function spaces and test spaces

LFAToolkit.gaussquadrature — Function

gaussquadrature(q)

Construct a Gauss-Legendre quadrature

quadraturepoints, quadratureweights = LFAToolkit.gaussquadrature(5);

# verify
truepoints = [
    -sqrt(5 + 2*sqrt(10/7))/3,
    -sqrt(5 - 2*sqrt(10/7))/3,
    0.0,
    sqrt(5 - 2*sqrt(10/7))/3,
    sqrt(5 + 2*sqrt(10/7))/3
];
trueweights = [
    (322-13*sqrt(70))/900,
    (322+13*sqrt(70))/900,
    128/225,
    (322+13*sqrt(70))/900,
    (322-13*sqrt(70))/900
];

diff = truepoints - quadraturepoints;
if abs(max(diff...)) > 1e-15
    println("Incorrect quadrature points");
end

diff = trueweights - quadratureweights;
if abs(abs(max(diff...))) > 1e-15
    println("Incorrect quadrature weights");
end

# output

LFAToolkit.lobattoquadrature — Function

lobattoquadrature(q, weights)

Construct a Gauss-Lobatto quadrature

quadraturepoints = LFAToolkit.lobattoquadrature(5, false);

# verify
truepoints = [-1.0, -sqrt(3/7), 0.0, sqrt(3/7), 1.0];

diff = truepoints - quadraturepoints;
if abs(max(diff...)) > 1e-15
    println("Incorrect quadrature points");
end

# output

quadraturepoints, quadratureweights = LFAToolkit.lobattoquadrature(5, true);

# verify
truepoints = [-1.0, -sqrt(3/7), 0.0, sqrt(3/7), 1.0];
trueweights = [1/10, 49/90, 32/45, 49/90, 1/10];

diff = truepoints - quadraturepoints;
if abs(max(diff...)) > 1e-15
    println("Incorrect quadrature points");
end

diff = trueweights - quadratureweights;
if abs(abs(max(diff...))) > 1e-15
    println("Incorrect quadrature weights");
end

# output

LFAToolkit.getnumbernodes — Function

getnumbernodes(basis)

Get the number of nodes for the basis

basis = TensorH1LagrangeBasis(4, 3, 2);
numbernodes = LFAToolkit.getnumbernodes(basis);

# verify
if numbernodes != 4^2
    println("Incorrect number of nodes")
end

# output

LFAToolkit.getnumberquadraturepoints — Function

getnumberquadraturepoints(basis)

Get the number of quadrature points for the basis

basis = TensorH1LagrangeBasis(4, 3, 2);
quadraturepoints = LFAToolkit.getnumberquadraturepoints(basis);
    
# verify
if quadraturepoints != 3^2
    println("Incorrect number of quadrature points")
end
    
# output

LFAToolkit.getinterpolation — Function

getinterpolation(basis)

Get full interpolation matrix for basis

for dimension in 1:3
    # get basis interpolation matrix
    basis = TensorH1LagrangeBasis(4, 3, dimension);
    interpolation = LFAToolkit.getinterpolation(basis);

    # verify
    for i in 1:3^dimension
        total = sum(interpolation[i, :]);
        if abs(total - 1.0) > 1e-15
            println("Incorrect interpolation matrix")
        end
    end
end

# output

LFAToolkit.getgradient — Function

getgradient(basis)

Get full gradient matrix for basis

for dimension in 1:3
    # get basis gradient matrix
    basis = TensorH1LagrangeBasis(4, 3, 3);
    gradient = LFAToolkit.getgradient(basis);

    # verify
    for i in 1:3*3^dimension
        total = sum(gradient[i, :]);
        if abs(total) > 1e-14
            println("Incorrect gradient matrix")
        end
    end
end

# output

LFAToolkit.getquadratureweights — Function

getquadratureweights(basis)

Get full quadrature weights vector for basis

for dimension in 1:3
    # get basis quadrature weights
    basis = TensorH1LagrangeBasis(4, 3, dimension);
    quadratureweights = LFAToolkit.getquadratureweights(basis);

    # verify
    trueweights1d = [5/9, 8/9, 5/9];
    trueweights = [];
    if dimension == 1
        trueweights = trueweights1d;
    elseif dimension == 2
        trueweights = kron(trueweights1d, trueweights1d);
    elseif dimension == 3
        trueweights = kron(trueweights1d, trueweights1d, trueweights1d);
    end

    diff = trueweights - quadratureweights;
    if abs(abs(max(diff...))) > 1e-15
        println("Incorrect quadrature weights");
    end
end
    
# output

Operator

LFAToolkit.getstencil — Function

getstencil(operator)

Compute or retrieve the stencil of operator for computing the symbol

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(4, 4, 2);
    
function massweakform(u::Array{Float64,1}, w::Array{Float64})
    v = u * w[1]
    return [v]
end
    
# mass operator
inputs = [
    OperatorField(basis, EvaluationMode.interpolation),
    OperatorField(basis, EvaluationMode.quadratureweights),
];
outputs = [OperatorField(basis, EvaluationMode.interpolation)];
mass = Operator(massweakform, mesh, inputs, outputs);
    
# stencil computation
stencil = LFAToolkit.getstencil(mass);

# verify
u = ones(4*4);
v = stencil * u;
    
total = sum(v);
if abs(total - 4.0) > 1e-14
    println("Incorrect mass matrix");
end

# test caching
stencil = LFAToolkit.getstencil(mass)
v = stencil * u;
    
total = sum(v);
if abs(total - 4.0) > 1e-14
    println("Incorrect mass matrix");
end
    
# output
    

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(4, 4, 2);
    
function diffusionweakform(du::Array{Float64,1}, w::Array{Float64})
    dv = du * w[1]
    return [dv]
end
    
# diffusion operator
inputs = [
    OperatorField(basis, EvaluationMode.gradient),
    OperatorField(basis, EvaluationMode.quadratureweights),
];
outputs = [OperatorField(basis, EvaluationMode.gradient)];
diffusion = Operator(diffusionweakform, mesh, inputs, outputs);
    
# stencil computation
stencil = LFAToolkit.getstencil(diffusion);
    
# verify
u = ones(4*4);
v = stencil * u;
    
total = sum(v);
if abs(total) > 1e-14
    println("Incorrect diffusion matrix");
end
    
# output