Finite Element Operator

LFAToolkit.getdimension — Function

getdimension(operator)

Retrieve the dimension of an operator

Arguments:

operator: operator to retrieve the dimension

Returns:

Dimension of the operator

Example

# setup
mesh = Mesh2D(1.0, 1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# verify
@assert mass.dimension == 2

# output

LFAToolkit.getelementmatrix — Function

getelementmatrix(operator)

Compute or retrieve the element matrix of operator for computing the symbol

Arguments:

operator: operator to compute element matrix

Returns:

Assembled element matrix

Mass matrix example:

# setup
mesh = Mesh2D(1.0, 1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# element matrix computation
# note: either syntax works
elementmatrix = LFAToolkit.getelementmatrix(mass);
elementmatrix = mass.elementmatrix;

# verify
u = ones(4*4);
v = elementmatrix*u;

total = sum(v);
@assert total ≈ 1.0

# output

Diffusion matrix example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 4, 4, mesh);

# element matrix computation
# note: either syntax works
elementmatrix = LFAToolkit.getelementmatrix(diffusion);
elementmatrix = diffusion.elementmatrix;

# verify
u = ones(4*4);
v = elementmatrix*u;

total = sum(v);
@assert abs(total) < 1e-14

# output

LFAToolkit.getmultiplicity — Function

getmultiplicity(operator)

Compute or retrieve the vector of node multiplicity for the operator

Arguments:

operator: operator to compute node multiplicity

Returns:

Vector of node multiplicity for the operator

Example

```jldoctest for dimension in 1:3 # setup mesh = [] if dimension == 1 mesh = Mesh1D(1.0); elseif dimension == 2 mesh = Mesh2D(1.0, 1.0); elseif dimension == 3 mesh = Mesh3D(1.0, 1.0, 1.0); end diffusion = GalleryOperator("diffusion", 5, 5, mesh);

# compute multiplicity
multiplicity = diffusion.multiplicity;

# verify
multiplicity1D = [2. 1. 1. 1. 2.]';
if dimension == 1
    @assert multiplicity ≈ multiplicity1D
elseif dimension == 2
    @assert multiplicity ≈ kron(multiplicity1D, multiplicity1D)
elseif dimension == 3
    @assert multiplicity ≈ kron(multiplicity1D, multiplicity1D, multiplicity1D)
end

end

LFAToolkit.getrowmodemap — Method

getrowmodemap(operator)

Compute or retrieve the matrix mapping the rows of the element matrix to the symbol matrix

Arguments:

operator: operator to compute row mode map

Returns:

Matrix mapping rows of element matrix to symbol matrix

Example:

# setup
mesh = Mesh1D(1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# note: either syntax works
modemap = LFAToolkit.getrowmodemap(mass);
modemap = mass.rowmodemap;

# verify
@assert modemap ≈ [1 0 0 1; 0 1 0 0; 0 0 1 0]

# output

LFAToolkit.getcolumnmodemap — Method

getcolumnmodemap(operator)

Compute or retrieve the matrix mapping the columns of the element matrix to the symbol matrix

Arguments:

operator: operator to compute column mode map

Returns:

Matrix mapping columns of element matrix to symbol matrix

Example:

# setup
mesh = Mesh1D(1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# note: either syntax works
modemap = LFAToolkit.getcolumnmodemap(mass);
modemap = mass.columnmodemap;

# verify
@assert modemap ≈ [1 0 0; 0 1 0; 0 0 1; 1 0 0]

# output

LFAToolkit.getinputcoordinates — Function

getinputcoordinates(operator)

Compute or retrieve the array of input coordinates

Arguments:

operator: operator to get input coordinates

Returns:

Array of input coordinates

LFAToolkit.getoutputcoordinates — Function

getoutputcoordinates(operator)

Compute or retrieve the array of output coordinates

Arguments:

operator: operator to get output coordinates

Returns:

Array of output coordinates

LFAToolkit.getnodecoordinatedifferences — Method

getnodecoordinatedifferences(operator)

Compute or retrieve the array of differences in coordinates between nodes

Arguments:

operator: operator to compute differences between node coordinates

Returns:

Array of differences in coordinates between nodes

Example:

using FastGaussQuadrature

# setup
mesh = Mesh1D(1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# note: either syntax works
nodedifferences = LFAToolkit.getnodecoordinatedifferences(mass);
nodedifferences = mass.nodecoordinatedifferences;

# verify
truenodes, = gausslobatto(4);
truenodedifferences = [
    (truenodes[j] - truenodes[i])/2.0 for i in 1:4, j in 1:4
];
@assert nodedifferences ≈ truenodedifferences

# output

LFAToolkit.massoperator — Function

massoperator(basis, mesh; parameters)

Convenience constructor for mass operator

Weak form:

$\int v u$

Arguments:

basis: basis for all operator fields to use

mesh: mesh for operator

Keyword Arguments:

parameters = nothing: named tuple of model parameters default: nothing, no parameters

Returns:

Mass matrix operator with basis on mesh

Example:

# mass operator
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(3, 4, 1, mesh.dimension);
mass = LFAToolkit.massoperator(basis, mesh);

# verify
println(mass)

# output
finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    interpolation
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    interpolation

LFAToolkit.diffusionoperator — Function

diffusionoperator(basis, mesh; parameters)

Convenience constructor for diffusion operator

Weak form:

$\int \nabla v \nabla u$

Arguments:

basis: basis for all operator fields to use

mesh: mesh for operator

Keyword Arguments:

parameters = nothing: named tuple of model parameters default: nothing, no parameters

Returns:

Diffusion operator with basis on mesh

Example:

# diffusion operator
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(3, 4, 1, mesh.dimension);
diffusion = LFAToolkit.diffusionoperator(basis, mesh);

# verify
println(diffusion)

# output

finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    gradient
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    gradient

LFAToolkit.advectionoperator — Function

advectionoperator(basis, mesh; parameters)

Convenience constructor for advection operator

Weak form:

$\int \nabla v u$

Arguments:

basis: basis for all operator fields to use

mesh: mesh for operator

Keyword Arguments:

parameters = ([wind = [1., 1.],): named tuple of model parameters, defines wind speed default: wind speed = [1.0, 1.0]

Returns:

Advection operator with basis on mesh

Example:

# advection operator
mesh = Mesh2D(1.0, 1.0);
mapping = hale_trefethen_strip_transformation(1.4);
basis = TensorH1LagrangeBasis(3, 4, 1, mesh.dimension, collocatedquadrature = false, mapping = mapping);
parameters = (wind = [1., 1.],);
advection = LFAToolkit.advectionoperator(basis, mesh; parameters = parameters);

# verify
println(advection)

# output

finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    interpolation
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    gradient

LFAToolkit.supgadvectionoperator — Function

supgadvectionoperator(basis, mesh; parameters)

Convenience constructor for SUPG advection operator

Weak form: Right hand side

$\int wind \nabla v u - wind wind τ \nabla v \nabla u$

Arguments:

basis: basis for all operator fields to use

mesh: mesh for operator

Keyword Arguments:

parameters = ([wind = [1., 1.], τ = 1.0): named tuple of model parameters, defines wind speed and SUPG scaling default: wind speed = [1.0, 1.0], τ = 1.0

Returns:

SUPG advection operator with basis on mesh

Example:

# supg advection operator
mesh = Mesh2D(1.0, 1.0);
mapping = nothing;
basis = TensorH1LagrangeBasis(3, 4, 1, mesh.dimension, collocatedquadrature = false, mapping = mapping);
parameters = (wind = [1., 1.], τ = 1.0);
supgadvection = LFAToolkit.supgadvectionoperator(basis, mesh; parameters = parameters);

# verify
println(supgadvection)

# output

finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation modes:
    interpolation
    gradient
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    gradient

LFAToolkit.supgmassoperator — Function

supgmassoperator(basis, mesh; parameters)

Convenience constructor for SUPG mass matrix operator

Weak form: Left hand side

$\int v u_t + wind τ u_t \nabla v$

Arguments:

basis: basis for all operator fields to use

mesh: mesh for operator

Keyword Arguments:

parameters = ([wind = [1., 1.], τ = 1.0): named tuple of model parameters, defines wind speed and SUPG scaling default: wind speed = [1.0, 1.0], τ = 1.0

Returns:

SUPG mass matrix operator with basis on mesh

Example:

# supg mass matrix operator
mesh = Mesh2D(1.0, 1.0);
mapping = nothing;
basis = TensorH1LagrangeBasis(3, 4, 1, mesh.dimension, collocatedquadrature = false, mapping = mapping);
parameters = (wind = [1., 1.], τ = 1.0);
supgmass = LFAToolkit.supgmassoperator(basis, mesh; parameters = parameters);

# verify
println(supgmass)

# output

finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    interpolation
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation modes:
    interpolation
    gradient