Preconditioner: BDDC

LFAToolkit.BDDC — Type

BDDC(fineoperator, coarseoperator, smoother, prolongation)

BDDC preconditioner for finite element operators

Arguments:

fineoperator: finite element operator to precondition

injectiontype: type of injection into subassembled space to use

Returns:

BDDC preconditioner object

LFAToolkit.getprimalvertices — Method

getprimalvertices(bddc)

Compute or retrieve the primal vertices for a BDDC preconditioner

Returns:

Vector of primal vertices for BDDC preconditioner

Example:

# setup
mesh = Mesh1D(1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
primalvertices = LFAToolkit.getprimalvertices(bddc);
primalvertices = bddc.primalvertices;

# verify
@assert primalvertices == [1, p]

# output

LFAToolkit.getsubassembledvertices — Method

getsubassembledvertices(bddc)

Compute or retrieve the subassembled vertices for a BDDC preconditioner

Returns:

Vector of subassembled vertices for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
subassembledvertices = LFAToolkit.getsubassembledvertices(bddc);
subassembledvertices = bddc.subassembledvertices;

# verify
@assert subassembledvertices == setdiff(1:p^2, [1, p, p^2-p+1, p^2])

# output

LFAToolkit.getinterfacevertices — Method

getinterfacevertices(bddc)

Compute or retrieve the interface vertices for a BDDC preconditioner

Returns:

Vector of interface vertices for BDDC preconditioner

Example:

# setup
mesh = Mesh1D(1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
interfacevertices = LFAToolkit.getinterfacevertices(bddc);
interfacevertices = bddc.interfacevertices;

# verify
@assert interfacevertices == [1, p]

# output

LFAToolkit.getinteriorvertices — Method

getinteriorvertices(bddc)

Compute or retrieve the interior vertices for a BDDC preconditioner

Returns:

Vector of interior vertices for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
interiorvertices = LFAToolkit.getinteriorvertices(bddc);
interiorvertices = bddc.interiorvertices;

# verify
trueinterfacevertices = [1:p..., p^2-p+1:p^2...]
for i = 1:p-2
    push!(trueinterfacevertices, i*p+1)
    push!(trueinterfacevertices, (i+1)*p)
end
trueinterfacevertices = sort(trueinterfacevertices)
@assert interiorvertices == setdiff(1:p^2, trueinterfacevertices)

# output

LFAToolkit.getsubassembledinverse — Function

getsubassembledinverse(bddc)

Compute or retrieve the subassembled subdomain solver for a BDDC preconditioner

Returns:

Matrix of inverse of subassembled subdomain matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
subassembledinverse = LFAToolkit.getsubassembledinverse(bddc);
subassembledinverse = bddc.subassembledinverse;

# verify
@assert subassembledinverse ≈ Matrix(diffusion.elementmatrix[bddc.subassembledvertices, bddc.subassembledvertices])^-1

# output

LFAToolkit.getinteriorinverse — Function

getinteriorinverse(bddc)

Compute or retrieve the interior solver for a BDDC preconditioner

Returns:

Matrix of inverse of interior matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
interiorinverse = LFAToolkit.getinteriorinverse(bddc);
interiorinverse = bddc.interiorinverse;

# verify
@assert interiorinverse ≈ Matrix(diffusion.elementmatrix[bddc.interiorvertices, bddc.interiorvertices])^-1

# output

LFAToolkit.getschur — Function

getschur(bddc)

Compute or retrieve the Schur complement matrix for a BDDC preconditioner

Returns:

Matrix for Schur complement

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
schur = LFAToolkit.getschur(bddc);
schur = bddc.schur;

# verify
@assert schur ≈ diffusion.elementmatrix[bddc.primalvertices, bddc.primalvertices] -
    diffusion.elementmatrix[bddc.primalvertices, bddc.subassembledvertices] *
    bddc.subassembledinverse *
    diffusion.elementmatrix[bddc.subassembledvertices, bddc.primalvertices]

# output

Missing docstring.

Missing docstring for LFAToolkit.getprimalsubassembled. Check Documenter's build log for details.

LFAToolkit.getinjection — Function

getcolumnmodemap(bddc)

Compute or retrieve the injection operator for the mixed BDDC matrix

Returns:

Matrix providing the injection operator of mixed BDDC matrix

Example:

LFAToolkit.getrowmodemap — Method

getrowmodemap(bddc)

Compute or retrieve the matrix mapping the rows of the mixed BDDC matrix to the symbol matrix

Returns:

Matrix mapping rows of mxied BDDC matrix to symbol matrix

Example:

# setup
mesh = Mesh1D(1.0);
finediffusion = GalleryOperator("diffusion", 4, 4, mesh);
bddc = LumpedBDDC(finediffusion);

# note: either syntax works
modemap = LFAToolkit.getrowmodemap(bddc);
modemap = bddc.rowmodemap;

# verify
@assert modemap ≈ [1 0 0; 0 1 0; 0 0 1]

# output

LFAToolkit.getcolumnmodemap — Method

getcolumnmodemap(bddc)

Compute or retrieve the matrix mapping the columns of the mixed BDDC matrix to the symbol matrix

Returns:

Matrix mapping columns of mixed BDDC matrix to symbol matrix

Example:

# setup
mesh = Mesh1D(1.0);
finediffusion = GalleryOperator("diffusion", 4, 4, mesh);
bddc = LumpedBDDC(finediffusion);

# note: either syntax works
modemap = LFAToolkit.getcolumnmodemap(bddc);
modemap = bddc.columnmodemap;

# verify
@assert modemap ≈ [1 0 0; 0 1 0; 0 0 1]

# output