Preconditioner: Chebyshev

LFAToolkit.getoperatordiagonalinverse — Method

getoperatordiagonalinverse(preconditioner)

Compute or retrieve the inverse of the symbol matrix diagonal for a Chebyshev preconditioner

Arguments:

preconditioner: preconditioner to compute diagonal inverse

Returns:

Symbol matrix diagonal inverse for the operator

Example:

# setup
mesh = Mesh1D(1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);

# preconditioner
chebyshev = Chebyshev(diffusion)

# note: either syntax works
diagonalinverse = LFAToolkit.getoperatordiagonalinverse(chebyshev);
diagonalinverse = chebyshev.operatordiagonalinverse;

# verify
@assert diagonalinverse ≈ [6/7 0; 0 3/4]
 
# output

LFAToolkit.geteigenvalueestimates — Method

geteigenvalueestimates(preconditioner)

Compute or retrieve the eigenvalue estimates for a Chebyshev preconditioner

Arguments:

preconditioner: preconditioner to compute eigenvalue estimates

Returns:

Eigenvalue estimates for the operator

Example:

# setup
mesh = Mesh1D(1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);

# preconditioner
chebyshev = Chebyshev(diffusion)

# estimate eigenvalues
eigenvalueestimates = LFAToolkit.geteigenvalueestimates(chebyshev);

# verify
@assert eigenvalueestimates ≈ [0, 15/7]
 
# output