Public API

Mesh

LFAToolkit.Mesh1D — Type

Mesh1D()

mesh = Mesh1D(1.0);

if (mesh.dx - 1.0) > 1e-15
    println("Incorrect x scale")
end

# output

LFAToolkit.Mesh2D — Type

Mesh2D()

mesh = Mesh2D(1.0, 0.5);

if (mesh.dx - 1.0) > 1e-15
    println("Incorrect x scale")
end
if (mesh.dy - 0.5) > 1e-15
    println("Incorrect y scale")
end

# output

LFAToolkit.Mesh3D — Type

Mesh2D()

mesh = Mesh3D(1.0, 0.5, 0.25);

if (mesh.dx - 1.0) > 1e-15
    println("Incorrect x scale")
end
if (mesh.dy - 0.5) > 1e-15
    println("Incorrect y scale")
end
if (mesh.dz - 0.25) > 1e-15
    printlt("Incorrect z scale")
end

# output

Basis

LFAToolkit.TensorBasis — Type

Tensor product basis

LFAToolkit.NonTensorBasis — Type

Non-tensor basis

LFAToolkit.TensorH1LagrangeBasis — Function

TensorH1LagrangeBasis()

Tensor product basis on Gauss-Lobatto points with Gauss-Legendre quadrature

basis = TensorH1LagrangeBasis(4, 3, 2);

# verify
if basis.p1d != 4
    println("Incorrect P1d");
end
if basis.q1d != 3
    println("Incorrect Q1d");
end
if basis.dimension != 2
    println("Incorrect dimension");
end

# output

Basis Evaluation Mode

LFAToolkit.EvaluationMode.EvalMode — Type

Basis evaluation mode for operator inputs and outputs

Interpolation - values interpolated to quadrature points

Gradient - derivatives evaluated at quadrature points

Quadrature Weights - quadrature weights

Operator Field

LFAToolkit.OperatorField — Type

Finite Element operator input or output, with a basis and evaluation mode

Operator

LFAToolkit.Operator — Type

Finite element operator comprising of a weak form and bases