Basis Evaluation Mode

This enum lists the supported basis evaluation modes for finite element operator fields. Note that the quadrature weights evaluation mode can only be used in an input field.

LFAToolkit.EvaluationMode.EvalMode — Type

Basis evaluation mode for operator inputs and outputs

Modes:

interpolation: values interpolated to quadrature points

gradient: derivatives evaluated at quadrature points

quadratureweights: quadrature weights

Example:

EvaluationMode.EvalMode

# output
Enum LFAToolkit.EvaluationMode.EvalMode:
interpolation = 0
gradient = 1
quadratureweights = 2