Rectangular Mesh

LFA is conducted on a infinite uniform grid. The mesh object allows the user to specify different scaling between the different dimensions of the mesh. This facilitates investigation of the smoothing and multigrid techniques on highly distorted meshes, such as boundary layer elements.

LFAToolkit.Mesh1D — Type

Mesh1D(dx)

One dimensional regular background mesh

Arguments:

dx: deformation in x dimension

Returns:

One dimensional mesh object

Example:

# generate 1D mesh
mesh = Mesh1D(1.0);

# verify
println(mesh)

# output
1d mesh:
    dx: 1.0

LFAToolkit.Mesh2D — Type

Mesh2D(dx, dy)

Two dimensional regular background mesh

Arguments:

dx: deformation in x dimension

dy: deformation in y dimension

Returns:

Two dimensional mesh object

Example:

# generate 2D mesh
mesh = Mesh2D(1.0, 0.5);

# verify
println(mesh)

# output
2d mesh:
    dx: 1.0
    dy: 0.5

LFAToolkit.Mesh3D — Type

Mesh3D(dx, dy, dz)

Three dimensional regular background mesh

Arguments:

dx: deformation in x dimension

dy: deformation in y dimension

dz: deformation in z dimension

Returns:

Three dimensional mesh object

Example:

# generate 3D mesh
mesh = Mesh3D(1.0, 0.5, 0.3);

# verify
println(mesh)

# output
3d mesh:
    dx: 1.0
    dy: 0.5
    dz: 0.3