Operator Field

LFAToolkit.OperatorField — Type

OperatorField(
    basis,
    evaluationmodes
)

Finite Element operator input or output, with a basis and evaluation mode

Arguments:

basis: finite element basis for the field

evaluationmodes: array of basis evaluation modes, note that quadrature weights must be listed in a separate operator field

Returns:

Finite element operator field object

Example:

# basis
basis = TensorH1LagrangeBasis(4, 3, 2);

# quadrature weights field
weightsfield = OperatorField(basis, [EvaluationMode.quadratureweights]);

# verify
println(weightsfield)

# input or output field
inputfield = OperatorField(basis, [
    EvaluationMode.interpolation,
    EvaluationMode.gradient,
]);

# verify
println(inputfield)

# output
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 3
    dimension: 2
  evaluation mode:
    quadratureweights
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 3
    dimension: 2
  evaluation modes:
    interpolation
    gradient