Preconditioner: Identity

The identity preconditioner object allows investigation of multigrid methods without smoothers.

Documentation

LFAToolkit.IdentityPC — Type

IdentityPC()

Identity preconditioner to investigate multigrid without smoother

Returns:

Identity preconditioner object

Example:

# preconditioner
identity = IdentityPC();

# verify
println(identity)

# output
identity preconditioner

LFAToolkit.computesymbols — Method

computesymbols(preconditioner, ω, θ)

Compute or retrieve the symbol matrix for a Jacobi preconditioned operator

Arguments:

preconditioner: Identity preconditioner to compute symbol matrix for

ω: Smoothing weighting factor array

θ: Fourier mode frequency array (one frequency per dimension)

Returns:

Symbol matrix for the identity preconditioner (I)

Example:

using LinearAlgebra

# preconditioner
identity = IdentityPC();

# compute symbols
A = computesymbols(identity, [], []);

# verify
@assert A ≈ I

# output