Preconditioner: Chebyshev

Example

This is an example of a Chebyshev smoother.

# ------------------------------------------------------------------------------
# Chebyshev smoother example
# ------------------------------------------------------------------------------

# setup
mesh = Mesh2D(1.0, 1.0)
basis = TensorH1LagrangeBasis(4, 4, 2)

function diffusionweakform(du::Array{Float64}, w::Array{Float64})
    dv = du*w[1]
    return [dv]
end

# diffusion operator
inputs = [
    OperatorField(basis, [EvaluationMode.gradient]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
]
outputs = [OperatorField(basis, [EvaluationMode.gradient])]
diffusion = Operator(diffusionweakform, mesh, inputs, outputs)

# Chebyshev smoother
chebyshev = Chebyshev(diffusion)

# compute operator symbols
A = computesymbols(chebyshev, [3], [π, π])
eigenvalues = real(eigvals(A))

# ------------------------------------------------------------------------------

Documentation

LFAToolkit.Chebyshev — Type

Chebyshev(operator)

Chebyshev polynomial preconditioner for finite element operators

Arguments:

operator: finite element operator to precondition

Returns:

Chebyshev preconditioner object

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# preconditioner
chebyshev = Chebyshev(mass);

# verify
println(chebyshev)

# output
chebyshev preconditioner:
eigenvalue estimates:
  estimated minimum 0.0278
  estimated maximum 0.3611
estimate scaling:
  λ_min = a * estimated min + b * estimated max
  λ_max = c * estimated min + d * estimated max
  a = 0.0000
  b = 0.1000
  c = 0.0000
  d = 1.1000

LFAToolkit.seteigenvalueestimatescaling — Function

seteigenvalueestimatescaling(preconditioner, eigenvaluebounds)

Set the scaling of the eigenvalue estimates for a Chebyshev preconditioner

Arguments:

eigenvaluebounds: array of 4 scaling factors to use when setting λmin and λmax based on eigenvalue estimates

λmin = a * estimated min + b * estimated max λmax = c * estimated min + d * estimated max

Example:

# setup
mesh = Mesh1D(1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);

# preconditioner
chebyshev = Chebyshev(diffusion)

# set eigenvalue estimate scaling
seteigenvalueestimatescaling(chebyshev, [0.0, 1.0, 2.0, 3.0]);
println(chebyshev)
 
# output
chebyshev preconditioner:
eigenvalue estimates:
  estimated minimum 1.0000
  estimated maximum 1.1429
estimate scaling:
  λ_min = a * estimated min + b * estimated max
  λ_max = c * estimated min + d * estimated max
  a = 0.0000
  b = 1.0000
  c = 2.0000
  d = 3.0000

LFAToolkit.computesymbols — Method

computesymbols(preconditioner, ω, θ)

Compute or retrieve the symbol matrix for a Chebyshev preconditioned operator

Arguments:

preconditioner: Chebyshev preconditioner to compute symbol matrix for

ω: Smoothing parameter array [degree], [degree, λmax], or [degree, λmin, λ_max]

θ: Fourier mode frequency array (one frequency per dimension)

Returns:

Symbol matrix for the Chebyshev preconditioned operator

Example:

using LinearAlgebra

for dimension in 1:3
    # setup
    mesh = []
    if dimension == 1
        mesh = Mesh1D(1.0);
    elseif dimension == 2
        mesh = Mesh2D(1.0, 1.0);
    elseif dimension == 3
        mesh = Mesh3D(1.0, 1.0, 1.0);
    end
    diffusion = GalleryOperator("diffusion", 3, 3, mesh);

    # preconditioner
    chebyshev = Chebyshev(diffusion);

    # compute symbols
    A = computesymbols(chebyshev, [1], π*ones(dimension));

    # verify
    using LinearAlgebra;
    eigenvalues = real(eigvals(A));
    if dimension == 1
        @assert max(eigenvalues...) ≈ 1/7
    elseif dimension == 2
        @assert min(eigenvalues...) ≈ -1/14
    elseif dimension == 3
        @assert min(eigenvalues...) ≈ -0.33928571428571486
    end
end

# output