Preconditioner: P-Multigrid

Examples

This is an example of a simple p-multigrid preconditioner.

# ------------------------------------------------------------------------------
# p-multigrid example
# ------------------------------------------------------------------------------

# setup
mesh = Mesh2D(1.0, 1.0)
finep = 3
coarsep = 2
dimension = 2
finebasis = TensorH1LagrangeBasis(finep, finep, dimension)
coarsebasis = TensorH1LagrangeBasis(coarsep, finep, dimension)
ctofbasis = TensorH1LagrangeBasis(coarsep, finep, dimension, lagrangequadrature = true)

function diffusionweakform(du::Array{Float64}, w::Array{Float64})
    dv = du*w[1]
    return [dv]
end

# diffusion operators
finediffusion = GalleryOperator("diffusion", finep, finep, mesh)
coarsediffusion = GalleryOperator("diffusion", coarsep, finep, mesh)

# Jacobi smoother
jacobi = Jacobi(finediffusion)

# p-multigrid preconditioner
multigrid = PMultigrid(finediffusion, coarsediffusion, jacobi, [ctofbasis])

# compute operator symbols
A = computesymbols(multigrid, [0.636], [1, 1], [π, π])
eigenvalues = real(eigvals(A))

# ------------------------------------------------------------------------------

This is an example of a multilevel p-multigrid preconditioner.

# ------------------------------------------------------------------------------
# p-multigrid multilevel example
# ------------------------------------------------------------------------------

# setup
mesh = Mesh2D(1.0, 1.0)
finep = 5
midp = 3
coarsep = 2
dimension = 2
ctombasis = TensorH1LagrangeBasis(coarsep, midp, dimension, lagrangequadrature = true)
mtofbasis = TensorH1LagrangeBasis(midp, finep, dimension, lagrangequadrature = true)

# diffusion operators
finediffusion = GalleryOperator("diffusion", finep, finep, mesh)
middiffusion = GalleryOperator("diffusion", midp, finep, mesh)
coarsediffusion = GalleryOperator("diffusion", coarsep, finep, mesh)

# Jacobi smoothers
finejacobi = Jacobi(finediffusion)
midjacobi = Jacobi(middiffusion)

# p-multigrid preconditioner
midmultigrid = PMultigrid(middiffusion, coarsediffusion, midjacobi, [ctombasis])
multigrid = PMultigrid(finediffusion, midmultigrid, finejacobi, [mtofbasis])

# compute operator symbols
A = computesymbols(multigrid, [0.651], [1, 1], [π, π])
eigenvalues = real(eigvals(A))

# ------------------------------------------------------------------------------

Documentation

LFAToolkit.PMultigrid — Type

PMultigrid(fineoperator, coarseoperator, smoother, prolongation)

P-Multigrid preconditioner for finite element operators

Arguments:

fineoperator: finite element operator to precondition

coarseoperator: coarse grid representation of finite element operator to precondition

smoother: error relaxation operator, such as Jacobi

prolongationbases: element prolongation bases from coarse to fine grid

Returns:

P-multigrid preconditioner object

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
ctofbasis = TensorH1LagrangeBasis(3, 5, 2, lagrangequadrature=true);

# operators
finediffusion = GalleryOperator("diffusion", 5, 5, mesh);
coarsediffusion = GalleryOperator("diffusion", 3, 5, mesh);

# smoother
jacobi = Jacobi(finediffusion);

# preconditioner
multigrid = PMultigrid(finediffusion, coarsediffusion, jacobi, [ctofbasis]);

# verify
println(multigrid)
println(multigrid.fineoperator)

# output
p-multigrid preconditioner
finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
tensor product basis:
    numbernodes1d: 5
    numberquadraturepoints1d: 5
    dimension: 2
  evaluation mode:
    gradient
operator field:
tensor product basis:
    numbernodes1d: 5
    numberquadraturepoints1d: 5
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
tensor product basis:
    numbernodes1d: 5
    numberquadraturepoints1d: 5
    dimension: 2
  evaluation mode:
    gradient

LFAToolkit.computesymbols — Method

computesymbols(multigrid, p, v, θ)

Compute or retrieve the symbol matrix for a Jacobi preconditioned operator

Arguments:

multigrid: PMultigrid preconditioner to compute symbol matrix for

p: Smoothing paramater array

v: Number of pre and post smooths

θ: Fourier mode frequency array (one frequency per dimension)

Returns:

Symbol matrix for the p-multigrid preconditioned operator

Example:

using LinearAlgebra

for dimension in 1:3
    # setup
    mesh = []
    if dimension == 1
        mesh = Mesh1D(1.0);
    elseif dimension == 2
        mesh = Mesh2D(1.0, 1.0);
    elseif dimension == 3
        mesh = Mesh3D(1.0, 1.0, 1.0);
    end
    ctofbasis = TensorH1LagrangeBasis(3, 5, dimension, lagrangequadrature=true);

    # operators
    finediffusion = GalleryOperator("diffusion", 5, 5, mesh);
    coarsediffusion = GalleryOperator("diffusion", 3, 5, mesh);

    # smoother
    jacobi = Jacobi(finediffusion);

    # preconditioner
    multigrid = PMultigrid(finediffusion, coarsediffusion, jacobi, [ctofbasis]);

    # compute symbols
    A = computesymbols(multigrid, [1.0], [1, 1], π*ones(dimension));

    # verify
    eigenvalues = real(eigvals(A));
    if dimension == 1
       @assert max(eigenvalues...) ≈ 0.64
    elseif dimension == 2
       @assert max(eigenvalues...) ≈ 0.9082562365654528
    elseif dimension == 3
       @assert max(eigenvalues...) ≈ 1.4359882222222669
    end
end

# output