Utilities

These functions are utilities for common analysis.

LFAToolkit.computesymbolsoverrange — Method

computesymbols(operator, θ)

Compute the eigenvalues and eigenvectors of the symbol matrix for an operator over a range of θ from -π/2 to 3π/2

Arguments:

operator: Finite element operator to compute symbol matrices for

numbersteps1d: Number of values of θ to sample in each dimension Note: numbersteps1d^dimension symbol matrices will be computed

Returns:

Values of θ sampled

Eigenvalues of symbol matrix at θ sampled

Eigenvectors of symbol matrix at θ sampled

Example:

using LFAToolkit;

numbersteps1d = 5

for dimension in 1:3
    # setup
    mesh = []
    if dimension == 1
        mesh = Mesh1D(1.0);
    elseif dimension == 2
        mesh = Mesh2D(1.0, 1.0);
    elseif dimension == 3
        mesh = Mesh3D(1.0, 1.0, 1.0);
    end
    diffusion = GalleryOperator("diffusion", 3, 3, mesh);

    # compute symbols
    (_, eigenvalues, _) = computesymbolsoverrange(diffusion, numbersteps1d);

    # verify
    eigenvalues = real(eigenvalues);
    if dimension == 1
        @assert minimum(eigenvalues[4, :]) ≈ 1
        @assert maximum(eigenvalues[4, :]) ≈ 4/3
    elseif dimension == 2
        @assert minimum(eigenvalues[19, :]) ≈ 2/3
        @assert maximum(eigenvalues[19, :]) ≈ 64/45
    elseif dimension == 3
        @assert minimum(eigenvalues[94, :]) ≈ 1/3
        @assert maximum(eigenvalues[94, :]) ≈ 256/225
    end
end

# output

LFAToolkit.computesymbolsoverrange — Method

computesymbols(operator, θ)

Compute the eigenvalues and eigenvectors of the symbol matrix for an operator over a range of θ from -π/2 to 3π/2

Arguments:

preconditioner: Preconditioner to compute symbol matries for

ω: Smoothing parameter array

numbersteps1d: Number of values of θ to sample in each dimension Note: numbersteps1d^dimension symbol matrices will be computed

Returns:

Values of θ sampled

Eigenvalues of symbol matrix at θ sampled

Eigenvectors of symbol matrix at θ sampled

Example:

using LFAToolkit;

numbersteps1d = 5

for dimension in 1:3
    # setup
    mesh = []
    if dimension == 1
        mesh = Mesh1D(1.0);
    elseif dimension == 2
        mesh = Mesh2D(1.0, 1.0);
    elseif dimension == 3
        mesh = Mesh3D(1.0, 1.0, 1.0);
    end
    diffusion = GalleryOperator("diffusion", 3, 3, mesh);

    # preconditioner
    chebyshev = Chebyshev(diffusion);

    # compute symbols
    (_, eigenvalues, _) = computesymbolsoverrange(chebyshev, [1], numbersteps1d);

    # verify
    eigenvalues = real(eigenvalues);
    if dimension == 1
        @assert minimum(eigenvalues[4, :]) ≈ 0.15151515151515105
        @assert maximum(eigenvalues[4, :]) ≈ 0.27272727272727226
    elseif dimension == 2
        @assert minimum(eigenvalues[19, :]) ≈ -0.25495098334134725
        @assert maximum(eigenvalues[19, :]) ≈ -0.17128758445192374
    elseif dimension == 3
        @assert minimum(eigenvalues[94, :]) ≈ -0.8181818181818181
        @assert maximum(eigenvalues[94, :]) ≈ -0.357575757575757
    end
end

# output