LFAToolkit

Local Fourier Analysis for arbitrary order finite element type operators

LFAToolkit.Basis — Type

Finite element basis for function spaces and test spaces

LFAToolkit.EvaluationMode — Type

Basis evaluation mode for operator inputs and outputs

LFAToolkit.Operator — Type

Finite element operator comprising of a weak form and bases

LFAToolkit.OperatorField — Type

Finite Element operator input or output, with a basis and evaluation mode

LFAToolkit.NonTensorBasis — Method

Constructor for non-tensor basis

LFAToolkit.TensorBasis — Method

Constructor for tensor product basis

LFAToolkit.TensorH1LagrangeBasis — Method

TensorH1LagrangeBasis

LFAToolkit.gaussquadrature — Method

gaussquadtarute()

Construct a Gauss-Legendre quadrature

LFAToolkit.getstencil — Method

getstencil()

Compute or retreive the stencil of operator for computing the symbol

LFAToolkit.lobattoquadrature — Method

lobattoquadrature()

Construct a Gauss-Lobatto quadrature