LFAToolkit

Local Fourier Analysis for arbitrary order finite element type operators

LFAToolkit.Basis — Type

Finite element basis for function spaces and test spaces

LFAToolkit.EvaluationMode — Type

Basis evaluation mode for operator inputs and outputs

LFAToolkit.Operator — Type

Finite element operator comprising of a weak form and bases

LFAToolkit.OperatorField — Type

Finite Element operator input or output, with a basis and evaluation mode

LFAToolkit.getstencil — Method

getstencil()

Compute or retreive the stencil of operator for computing the symbol