LFAToolkit

Local Fourier Analysis for arbitrary order finite element type operators

LFAToolkit.Basis — Type

Finite element basis for function spaces and test spaces

LFAToolkit.Operator — Type

Finite element operator comprising of a weak form and bases

LFAToolkit.OperatorField — Type

Finite Element operator input or output, with a basis and evaluation mode

LFAToolkit.NonTensorBasis — Method

Constructor for non-tensor basis

LFAToolkit.TensorBasis — Method

Constructor for tensor product basis

LFAToolkit.TensorH1LagrangeBasis — Method

TensorH1LagrangeBasis()

Tensor product basis on Gauss-Lobatto points with Gauss-Legendre quadrature

```@meta
DocTestSetup = quote
    using LFAToolkit
end
```

```jldoctest
basis = TensorH1LagrangeBasis(4, 4, 1, 1)

# output
LFAToolkit.Basis(4, 4, 1, 1, true, [-1.0, -0.4472135954999579, 0.4472135954999579, 1.0], [-0.8611363115940526, -0.3399810435848563, 0.3399810435848563, 0.8611363115940526], [0.34785484513745374, 0.6521451548625461, 0.6521451548625461, 0.34785484513745374], [0.6299431661034454 0.472558747113818 -0.14950343104607952 0.04700151782881607; -0.07069479527385582 0.972976186258263 0.13253992624542693 -0.03482131722983419; -0.03482131722983419 0.13253992624542696 0.9729761862582628 -0.07069479527385582; 0.04700151782881607 -0.14950343104607955 0.47255874711381796 0.6299431661034455], [-2.341837415390958 2.787944890537088 -0.6351041115519563 0.18899663640582656; -0.5167021357255352 -0.48795249031352683 1.3379050992756671 -0.3332504732366054; 0.33325047323660545 -1.3379050992756674 0.4879524903135269 0.5167021357255351; -0.1889966364058266 0.6351041115519563 -2.7879448905370876 2.3418374153909585])
```

```@meta
DocTestSetup = nothing
```

LFAToolkit.gaussquadrature — Method

gaussquadrature()

Construct a Gauss-Legendre quadrature

```@meta
DocTestSetup = quote
  using LFAToolkit
end
```

```jldoctest
quadraturepoints = LFAToolkit.gaussquadrature(5)

# output
([-0.906179845938664, -0.5384693101056831, 1.232595164407831e-32, 0.5384693101056831, 0.906179845938664], [0.2369268850561885, 0.47862867049936647, 0.5688888888888889, 0.47862867049936647, 0.2369268850561885])
```

```@meta
DocTestSetup = nothing
```

LFAToolkit.getstencil — Method

getstencil()

Compute or retreive the stencil of operator for computing the symbol

LFAToolkit.lobattoquadrature — Method

lobattoquadrature()

Construct a Gauss-Lobatto quadrature

```@meta
DocTestSetup = quote
  using LFAToolkit
end
```

```jldoctest
quadraturepoints = LFAToolkit.lobattoquadrature(5, false)

# output
5-element Array{Float64,1}:
-1.0
-0.6546536707079771
-1.530808498934193e-17
 0.6546536707079771
 1.0
```

```jldoctest
quadraturepoints, quadratureweights = LFAToolkit.lobattoquadrature(5, false)

# output
([-1.0, -0.6546536707079771, -1.530808498934193e-17, 0.6546536707079771, 1.0], [0.1, 0.5444444444444443, 0.7111111111111111, 0.5444444444444443, 0.1])
```

```@meta
DocTestSetup = nothing
```