LFAToolkit

Local Fourier Analysis for arbitrary order finite element type operators

LFAToolkit.Basis — Type

Finite element basis for function spaces and test spaces

LFAToolkit.Mesh — Type

Rectangular mesh with independent scaling in each dimesion

LFAToolkit.Mesh1D — Type

Mesh1D()

mesh = Mesh1D(1.0);

if (mesh.dx - 1.0) > 1e-15
    println("Incorrect x scale")
end

# output

LFAToolkit.Mesh2D — Type

Mesh2D()

mesh = Mesh2D(1.0, 0.5);

if (mesh.dx - 1.0) > 1e-15
    println("Incorrect x scale")
end
if (mesh.dy - 0.5) > 1e-15
    println("Incorrect y scale")
end

# output

LFAToolkit.Mesh3D — Type

Mesh2D()

mesh = Mesh3D(1.0, 0.5, 0.25);

if (mesh.dx - 1.0) > 1e-15
    println("Incorrect x scale")
end
if (mesh.dy - 0.5) > 1e-15
    println("Incorrect y scale")
end
if (mesh.dz - 0.25) > 1e-15
    printlt("Incorrect z scale")
end

# output

LFAToolkit.NonTensorBasis — Type

Non-tensor basis

LFAToolkit.Operator — Type

Finite element operator comprising of a weak form and bases

LFAToolkit.OperatorField — Type

Finite Element operator input or output, with a basis and evaluation mode

LFAToolkit.TensorBasis — Type

Tensor product basis

LFAToolkit.TensorH1LagrangeBasis — Method

TensorH1LagrangeBasis()

Tensor product basis on Gauss-Lobatto points with Gauss-Legendre quadrature

basis = TensorH1LagrangeBasis(4, 3, 2, 1);

# verify
if basis.p1d != 4
    println("Incorrect P1d");
end
if basis.q1d != 3
    println("Incorrect Q1d");
end
if basis.dimension != 2
    println("Incorrect dimension");
end
if basis.numbercomponents != 1
    println("Incorrect number of components");
end

# output

LFAToolkit.gaussquadrature — Method

gaussquadrature()

Construct a Gauss-Legendre quadrature

quadraturepoints, quadratureweights = LFAToolkit.gaussquadrature(5);

# verify
truepoints = [-sqrt(5 + 2*sqrt(10/7))/3, -sqrt(5 - 2*sqrt(10/7))/3, 0.0, sqrt(5 - 2*sqrt(10/7))/3, sqrt(5 + 2*sqrt(10/7))/3];
trueweights = [(322-13*sqrt(70))/900, (322+13*sqrt(70))/900, 128/225, (322+13*sqrt(70))/900, (322-13*sqrt(70))/900];

diff = truepoints - quadraturepoints;
if abs(max(diff...)) > 1e-15
    println("Incorrect quadrature points");
end

diff = trueweights - quadratureweights;
if abs(abs(max(diff...))) > 1e-15
    println("Incorrect quadrature weights");
end

# output

LFAToolkit.getgradient — Method

getgradient()

Get full gradient matrix for basis

basis = TensorH1LagrangeBasis(4, 3, 1, 1);
gradient = LFAToolkit.getgradient(basis);

# verify
for i in 1:3
    sum = 0.0
    for j = 1:4
        sum += gradient[i, j];
    end
    if abs(sum) > 1e-15
        println("Incorrect gradent matrix")
    end
end

# output

basis = TensorH1LagrangeBasis(4, 3, 2, 1);
gradient = LFAToolkit.getgradient(basis);

# verify
for i in 1:2*3^2
    sum = 0.0
    for j = 1:4^2
        sum += gradient[i, j];
    end
    if abs(sum) > 1e-15
        println("Incorrect gradient matrix")
    end
end

# output

basis = TensorH1LagrangeBasis(4, 3, 3, 1);
gradient = LFAToolkit.getgradient(basis);

# verify
for i in 1:3*3^3
    sum = 0.0
    for j = 1:4^3
        sum += gradient[i, j];
    end
    if abs(sum) > 1e-14
        println("Incorrect gradient matrix")
    end
end

# output

LFAToolkit.getinterpolation — Method

getinterpolation()

Get full interpolation matrix for basis

basis = TensorH1LagrangeBasis(4, 3, 1, 1);
interpolation = LFAToolkit.getinterpolation(basis);

# verify
for i in 1:3
    sum = 0.0
    for j = 1:4
        sum += interpolation[i, j];
    end
    if abs(sum - 1.0) > 1e-15
        println("Incorrect interpolation matrix")
    end
end

# output

basis = TensorH1LagrangeBasis(4, 3, 2, 1);
interpolation = LFAToolkit.getinterpolation(basis);

# verify
for i in 1:3^2
    sum = 0.0
    for j = 1:4^2
        sum += interpolation[i, j];
    end
    if abs(sum - 1.0) > 1e-15
        println("Incorrect interpolation matrix")
    end
end

# output

basis = TensorH1LagrangeBasis(4, 3, 3, 1);
interpolation = LFAToolkit.getinterpolation(basis);

# verify
for i in 1:3^3
    sum = 0.0
    for j = 1:4^3
        sum += interpolation[i, j];
    end
    if abs(sum - 1.0) > 1e-15
        println("Incorrect interpolation matrix")
    end
end

# output

LFAToolkit.getnumbernodes — Method

getnumbernodes()

Get the number of nodes for the basis

basis = TensorH1LagrangeBasis(4, 3, 2, 1);
numbernodes = LFAToolkit.getnumbernodes(basis);

# verify
if numbernodes != 16
    println("Incorrect number of nodes")
end

# output

LFAToolkit.getnumberquadraturepoints — Method

getnumberquadraturepoints()

Get the number of quadrature points for the basis

basis = TensorH1LagrangeBasis(4, 3, 2, 1);
quadraturepoints = LFAToolkit.getnumberquadraturepoints(basis);
    
# verify
if quadraturepoints != 9
    println("Incorrect number of quadrature points")
end
    
# output

LFAToolkit.getquadratureweights — Method

getquadratureweights()

Get full quadrature weights vector for basis

basis = TensorH1LagrangeBasis(4, 3, 1, 1);
quadratureweights = LFAToolkit.getquadratureweights(basis);

# verify
trueweights = [5/9, 8/9, 5/9];

diff = trueweights - quadratureweights;
if abs(abs(max(diff...))) > 1e-15
    println("Incorrect quadrature weights");
end

# output

basis = TensorH1LagrangeBasis(4, 3, 2, 1);
quadratureweights = LFAToolkit.getquadratureweights(basis);

# verify
trueweights1d = [5/9, 8/9, 5/9];
trueweights = kron(trueweights1d, trueweights1d);

diff = trueweights - quadratureweights;
if abs(abs(max(diff...))) > 1e-15
    println("Incorrect quadrature weights");
end
    
# output

basis = TensorH1LagrangeBasis(4, 3, 3, 1);
quadratureweights = LFAToolkit.getquadratureweights(basis);

# verify
trueweights1d = [5/9, 8/9, 5/9];
trueweights = kron(trueweights1d, trueweights1d, trueweights1d);

diff = trueweights - quadratureweights;
if abs(abs(max(diff...))) > 1e-15
    println("Incorrect quadrature weights");
end
    
# output

LFAToolkit.getstencil — Method

getstencil()

Compute or retrieve the stencil of operator for computing the symbol

LFAToolkit.lobattoquadrature — Method

lobattoquadrature()

Construct a Gauss-Lobatto quadrature

quadraturepoints = LFAToolkit.lobattoquadrature(5, false);

# verify
truepoints = [-1.0, -sqrt(3/7), 0.0, sqrt(3/7), 1.0];

diff = truepoints - quadraturepoints;
if abs(max(diff...)) > 1e-15
    println("Incorrect quadrature points");
end

# output

quadraturepoints, quadratureweights = LFAToolkit.lobattoquadrature(5, true);

# verify
truepoints = [-1.0, -sqrt(3/7), 0.0, sqrt(3/7), 1.0];
trueweights = [1/10, 49/90, 32/45, 49/90, 1/10];

diff = truepoints - quadraturepoints;
if abs(max(diff...)) > 1e-15
    println("Incorrect quadrature points");
end

diff = trueweights - quadratureweights;
if abs(abs(max(diff...))) > 1e-15
    println("Incorrect quadrature weights");
end

# output