Private API

This page documents the private API of the LFAToolkit.

Mesh

LFAToolkit.Mesh — Type

Rectangular mesh with independent scaling in each dimesion

Basis

LFAToolkit.Basis — Type

Finite element basis for function spaces and test spaces

LFAToolkit.gaussquadrature — Function

gaussquadrature(q)

Construct a Gauss-Legendre quadrature

Arguments:

q: number of Gauss-Legendre points

Returns:

Gauss-Legendre quadrature points and weights

Example:

# generate Gauss-Legendre points and weights
quadraturepoints, quadratureweights = LFAToolkit.gaussquadrature(5);

# verify
truepoints = [
    -sqrt(5 + 2*sqrt(10/7))/3,
    -sqrt(5 - 2*sqrt(10/7))/3,
    0.0,
    sqrt(5 - 2*sqrt(10/7))/3,
    sqrt(5 + 2*sqrt(10/7))/3
];
trueweights = [
    (322-13*sqrt(70))/900,
    (322+13*sqrt(70))/900,
    128/225,
    (322+13*sqrt(70))/900,
    (322-13*sqrt(70))/900
];

diff = truepoints - quadraturepoints;
@assert abs(max(diff...)) < 1e-15

diff = trueweights - quadratureweights;
@assert abs(max(diff...)) < 1e-15

# output

LFAToolkit.lobattoquadrature — Function

lobattoquadrature(q, weights)

Construct a Gauss-Lobatto quadrature

Arguments:

q: number of Gauss-Lobatto points

weights: boolean flag indicating if quadrature weights are desired

Returns:

Gauss-Lobatto quadrature points or points and weights

Example:

# generate Gauss-Lobatto points
quadraturepoints = LFAToolkit.lobattoquadrature(5, false);

# verify
truepoints = [-1.0, -sqrt(3/7), 0.0, sqrt(3/7), 1.0];

diff = truepoints - quadraturepoints;
@assert abs(max(diff...)) < 1e-15

# generate Gauss-Lobatto points and weights
quadraturepoints, quadratureweights = LFAToolkit.lobattoquadrature(5, true);

# verify
trueweights = [1/10, 49/90, 32/45, 49/90, 1/10];

diff = trueweights - quadratureweights;
@assert abs(max(diff...)) < 1e-15

# output

LFAToolkit.getnumbernodes — Function

getnumbernodes(basis)

Get the number of nodes for the basis

Arguments:

basis: basis to compute number of nodes

Returns:

Integer number of basis nodes

Example:

# get number of nodes for basis
basis = TensorH1LagrangeBasis(4, 3, 2);
numbernodes = LFAToolkit.getnumbernodes(basis);

# verify
@assert numbernodes == 4^2

# output

LFAToolkit.getnodes — Function

getnodes(basis)

Get nodes for basis

Returns:

Basis nodes array

Arguments:

basis: basis to compute nodes

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis quadrature weights
    basis = TensorH1LagrangeBasis(4, 3, dimension);
    nodes = LFAToolkit.getnodes(basis);

    # verify
    truenodes1d = [-1, -sqrt(1/5), sqrt(1/5), 1];
    truenodes = [];
    if dimension == 1
        truenodes = truenodes1d;
    elseif dimension == 2
        truenodes = transpose(hcat([[[x, y] for x in basis.nodes1d, y in basis.nodes1d]...]...));
    elseif dimension == 3
        truenodes = transpose(hcat([[
            [x, y, z] for x in basis.nodes1d, y in basis.nodes1d, z in basis.nodes1d
        ]...]...));
    end

    diff = truenodes - nodes;
    @assert abs(max(max(diff...)...)) < 1e-15
end
    
# output

LFAToolkit.getnumberquadraturepoints — Function

getnumberquadraturepoints(basis)

Get the number of quadrature points for the basis

Arguments:

basis: basis to compute number of quadrature points

Returns:

Integer number of basis quadrature points

Example:

# get number of quadrature points for basis
basis = TensorH1LagrangeBasis(4, 3, 2);
quadraturepoints = LFAToolkit.getnumberquadraturepoints(basis);
    
# verify
@assert quadraturepoints == 3^2
    
# output

LFAToolkit.getinterpolation — Function

getinterpolation(basis)

Get full interpolation matrix for basis

Arguments:

basis: basis to compute interpolation matrix

Returns:

Basis interpolation matrix

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis interpolation matrix
    basis = TensorH1LagrangeBasis(4, 3, dimension);
    interpolation = LFAToolkit.getinterpolation(basis);

    # verify
    for i in 1:3^dimension
        total = sum(interpolation[i, :]);
        @assert abs(total - 1.0) < 1e-15
    end
end

# output

LFAToolkit.getgradient — Function

getgradient(basis)

Get full gradient matrix for basis

Arguments:

basis: basis to compute gradient matrix

Returns:

Basis gradient matrix

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis gradient matrix
    basis = TensorH1LagrangeBasis(4, 3, dimension);
    gradient = LFAToolkit.getgradient(basis);

    # verify
    for i in 1:dimension*3^dimension
        total = sum(gradient[i, :]);
        @assert abs(total) < 1e-14
    end
end

# output

LFAToolkit.getquadratureweights — Function

getquadratureweights(basis)

Get full quadrature weights vector for basis

Returns:

Basis quadrature weights vector

Arguments:

basis: basis to compute quadrature weights

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis quadrature weights
    basis = TensorH1LagrangeBasis(4, 3, dimension);
    quadratureweights = LFAToolkit.getquadratureweights(basis);

    # verify
    trueweights1d = [5/9, 8/9, 5/9];
    trueweights = [];
    if dimension == 1
        trueweights = trueweights1d;
    elseif dimension == 2
        trueweights = kron(trueweights1d, trueweights1d);
    elseif dimension == 3
        trueweights = kron(trueweights1d, trueweights1d, trueweights1d);
    end

    diff = trueweights - quadratureweights;
    @assert abs(max(diff...)) < 1e-15
end
    
# output

Operator

LFAToolkit.getstencil — Function

getstencil(operator)

Compute or retrieve the stencil of operator for computing the symbol

Arguments:

operator: operator to compute element stencil

Returns:

Assembled element matrix

Mass matrix example:

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(4, 4, 2);
    
function massweakform(u::Array{Float64}, w::Array{Float64})
    v = u * w[1]
    return [v]
end
    
# mass operator
inputs = [
    OperatorField(basis, [EvaluationMode.interpolation]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.interpolation])];
mass = Operator(massweakform, mesh, inputs, outputs);
    
# stencil computation
stencil = LFAToolkit.getstencil(mass);

# verify
u = ones(4*4);
v = stencil * u;
    
total = sum(v);
@assert abs(total - 4.0) < 1e-14

# test caching
stencil = LFAToolkit.getstencil(mass)
v = stencil * u;
    
total = sum(v);
@assert abs(total - 4.0) < 1e-14
    
# output

Diffusion matrix example:

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(4, 4, 2);
    
function diffusionweakform(du::Array{Float64}, w::Array{Float64})
    dv = du * w[1]
    return [dv]
end
    
# diffusion operator
inputs = [
    OperatorField(basis, [EvaluationMode.gradient]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.gradient])];
diffusion = Operator(diffusionweakform, mesh, inputs, outputs);
    
# stencil computation
stencil = LFAToolkit.getstencil(diffusion);
    
# verify
u = ones(4*4);
v = stencil * u;
    
total = sum(v);
@assert abs(total) < 1e-14
    
# output