Finite Element Basis

LFAToolkit.AbstractBasis — Type

Finite element basis for function spaces and test spaces

LFAToolkit.gaussquadrature — Function

gaussquadrature(q)

Construct a Gauss-Legendre quadrature

Arguments:

q: number of Gauss-Legendre points

Returns:

Gauss-Legendre quadrature points and weights

Example:

# generate Gauss-Legendre points and weights
quadraturepoints, quadratureweights = LFAToolkit.gaussquadrature(5);

# verify
truepoints = [
    -√(5 + 2*√(10/7))/3,
    -√(5 - 2*√(10/7))/3,
    0.0,
    √(5 - 2*√(10/7))/3,
    √(5 + 2*√(10/7))/3
];
@assert truepoints ≈ quadraturepoints

trueweights = [
    (322-13*√70)/900,
    (322+13*√70)/900,
    128/225,
    (322+13*√70)/900,
    (322-13*√70)/900
];
@assert trueweights ≈ quadratureweights

# output

LFAToolkit.gausslobattoquadrature — Function

gausslobattoquadrature(q, weights)

Construct a Gauss-Legendre-Lobatto quadrature

Arguments:

q: number of Gauss-Legendre-Lobatto points

weights: boolean flag indicating if quadrature weights are desired

Returns:

Gauss-Legendre-Lobatto quadrature points or points and weights

Example:

# generate Gauss-Legendre-Lobatto points
quadraturepoints = LFAToolkit.gausslobattoquadrature(5, false);

# verify
truepoints = [-1.0, -√(3/7), 0.0, √(3/7), 1.0];
@assert truepoints ≈ quadraturepoints

# generate Gauss-Legendre-Lobatto points and weights
quadraturepoints, quadratureweights = LFAToolkit.gausslobattoquadrature(5, true);

# verify
trueweights = [1/10, 49/90, 32/45, 49/90, 1/10];
@assert trueweights ≈ quadratureweights

# output

LFAToolkit.TensorMacroElementBasisFrom1D — Function

TensorMacroElementBasisFrom1D(
    numbernodes1d,
    numberquadraturepoints1d,
    numbercomponents,
    dimension,
    numberelements1d,
    basis1dmicro,
    overlapquadraturepoints = false,
)

Tensor product macro-element basis from 1d single element tensor product basis

Arguments:

numbernodes1d: number of basis nodes in 1 dimension

numberquadraturepoints1d: number of quadrature points in 1 dimension

numbercomponents: number of components

dimension: dimension of basis

numberelements1d: number of elements in macro-element

basis1dmicro: 1d micro element basis to replicate

Keyword Arguments:

overlapquadraturepoints: Overlap quadrature points between elements, for prolongation default: false

Returns:

Tensor product macro-element basis object

LFAToolkit.TensorHProlongationBasis — Function

TensorHProlongationBasis(
    coarsenodes1d,
    finenodes1d,
    numbercomponents,
    dimension,
    numberfineelements1d,
)

Tensor product h-prolongation basis

Arguments:

coarsenodes1d: coarse grid node coordinates in 1 dimension

finenodes1d: fine grid node coordinates in 1 dimension

numbercomponents: number of components

dimension: dimension of basis

numberfineelements1d: number of fine grid elements

Returns:

H1 Lagrange tensor product h-prolongation basis object

LFAToolkit.buildinterpolationandgradient — Function

buildinterpolationandgradient(
    nodes,
    quadraturepoints,
)

Build one dimensional interpolation and gradient matrices, from Fornberg 1998

Arguments:

nodes: 1d basis nodes

quadraturepoints: 1d basis quadrature points

Returns:

One dimensional interpolation and gradient matrices

Example:

# get nodes, quadrature points, and weights
numbernodes = 3;
numberquadraturepoints = 4;
nodes = LFAToolkit.gausslobattoquadrature(numbernodes, false);
quadraturepoints, quadratureweights1d = LFAToolkit.gaussquadrature(numberquadraturepoints);

# build interpolation, gradient matrices
interpolation, gradient = LFAToolkit.buildinterpolationandgradient(nodes, quadraturepoints);

# verify
for i in 1:numberquadraturepoints
    total = sum(interpolation[i, :]);
    @assert total ≈ 1.0

    total = sum(gradient[i, :]);
    @assert abs(total) < 1e-14
end

# output

LFAToolkit.getnumbernodes — Function

getnumbernodes(basis)

Get the number of nodes for the basis

Arguments:

basis: basis to compute number of nodes

Returns:

Integer number of basis nodes

Example:

# get number of nodes for basis
basis = TensorH1LagrangeBasis(4, 3, 2, 2);

# note: either syntax works
numbernodes = LFAToolkit.getnumbernodes(basis);
numbernodes = basis.numbernodes;

# verify
@assert numbernodes == 4^2

# output

LFAToolkit.getnodes — Function

getnodes(basis)

Get nodes for basis

Returns:

Basis nodes array

Arguments:

basis: basis to compute nodes

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis quadrature weights
    basis = TensorH1LagrangeBasis(4, 3, 2, dimension);

    # note: either syntax works
    nodes = LFAToolkit.getnodes(basis);
    nodes = basis.nodes;

    # verify
    truenodes1d = [-1, -√(1/5), √(1/5), 1];
    truenodes = [];
    if dimension == 1
        truenodes = truenodes1d;
    elseif dimension == 2
        truenodes =
            transpose(hcat([[[x, y] for x in truenodes1d, y in truenodes1d]...]...));
    elseif dimension == 3
        truenodes = transpose(hcat([[
            [x, y, z] for x in truenodes1d, y in truenodes1d, z in truenodes1d
        ]...]...));
    end

    @assert truenodes ≈ nodes
end
    
# output

LFAToolkit.getnumberquadraturepoints — Function

getnumberquadraturepoints(basis)

Get the number of quadrature points for the basis

Arguments:

basis: basis to compute number of quadrature points

Returns:

Integer number of basis quadrature points

Example:

# get number of quadrature points for basis
basis = TensorH1LagrangeBasis(4, 3, 2, 2);

# note: either syntax works
numberquadraturepoints = LFAToolkit.getnumberquadraturepoints(basis);
numberquadraturepoints = basis.numberquadraturepoints;
    
# verify
@assert numberquadraturepoints == 3^2
    
# output

LFAToolkit.getquadraturepoints — Function

getquadraturepoints(basis)

Get quadrature points for basis

Returns:

Basis quadrature points array

Arguments:

basis: basis to compute quadrature points

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis quadrature weights
    basis = TensorH1LagrangeBasis(4, 3, 2, dimension);

    # note: either syntax works
    quadraturepoints = LFAToolkit.getquadraturepoints(basis);
    quadraturepoints = basis.quadraturepoints;

    # verify
    truequadraturepoints1d = [-√(3/5), 0, √(3/5)];
    truequadraturepoints = [];
    if dimension == 1
        truequadraturepoints = truequadraturepoints1d;
    elseif dimension == 2
        truequadraturepoints = transpose(hcat([[
            [x, y] for x in truequadraturepoints1d, y in truequadraturepoints1d
        ]...]...));
    elseif dimension == 3
        truequadraturepoints = transpose(hcat([[
            [x, y, z]
            for
            x in truequadraturepoints1d,
            y in truequadraturepoints1d, z in truequadraturepoints1d
        ]...]...));
    end

    @assert truequadraturepoints ≈ quadraturepoints
end
    
# output

LFAToolkit.getquadratureweights — Function

getquadratureweights(basis)

Get full quadrature weights vector for basis

Returns:

Basis quadrature weights vector

Arguments:

basis: basis to compute quadrature weights

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis quadrature weights
    basis = TensorH1LagrangeBasis(4, 3, 2, dimension);

    # note: either syntax works
    quadratureweights = LFAToolkit.getquadratureweights(basis);
    quadratureweights = basis.quadratureweights;

    # verify
    trueweights1d = [5/9, 8/9, 5/9];
    trueweights = [];
    if dimension == 1
        trueweights = trueweights1d;
    elseif dimension == 2
        trueweights = kron(trueweights1d, trueweights1d);
    elseif dimension == 3
        trueweights = kron(trueweights1d, trueweights1d, trueweights1d);
    end

    @assert trueweights ≈ quadratureweights
end
    
# output

LFAToolkit.getinterpolation — Function

getinterpolation(basis)

Get full interpolation matrix for basis

Arguments:

basis: basis to compute interpolation matrix

Returns:

Basis interpolation matrix

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis interpolation matrix
    basis = TensorH1LagrangeBasis(4, 3, 2, dimension);

    # note: either syntax works
    interpolation = LFAToolkit.getinterpolation(basis);
    interpolation = basis.interpolation;

    # verify
    for i in 1:3^dimension
        total = sum(interpolation[i, :]);
        @assert total ≈ 1.0
    end
end

# output

LFAToolkit.getgradient — Function

getgradient(basis)

Get full gradient matrix for basis

Arguments:

basis: basis to compute gradient matrix

Returns:

Basis gradient matrix

Example:

# test for all supported dimensions
for dimension in 1:3
    # get basis gradient matrix
    basis = TensorH1LagrangeBasis(4, 3, 2, dimension);

    # note: either syntax works
    gradient = LFAToolkit.getgradient(basis);
    gradient = basis.gradient;

    # verify
    for i in 1:dimension*3^dimension
        total = sum(gradient[i, :]);
        @assert abs(total) < 1e-14
    end
end

# output

LFAToolkit.getnumbermodes — Function

getnumbermodes(basis)

Get number of modes for basis

Arguments:

basis: basis to compute number of modes

Returns:

Number of modes for basis

Example:

# test for all supported dimensions
for dimension in 1:3
    # get number of basis modes
    basis = TensorH1LagrangeBasis(4, 3, 2, dimension);

    # note: either syntax works
    numbermodes = LFAToolkit.getnumbermodes(basis);
    numbermodes = basis.numbermodes;

    # verify
    @assert numbermodes == 2*3^dimension
end

# output

LFAToolkit.getmodemap — Function

getmodemap(basis)

Get mode mapping vector for basis

Arguments:

basis: basis to compute mode map vector

Returns:

Basis mode map vector

Example:

# test for all supported dimensions
for dimension in 1:3
    # get mode map vector
    basis = TensorH1LagrangeBasis(4, 3, 1, dimension);

    # note: either syntax works
    modemap = LFAToolkit.getmodemap(basis);
    modemap = basis.modemap;

    # verify
    truemodemap1d = [1, 2, 3, 1];
    truemodemap = [];
    if dimension == 1
        truemodemap = truemodemap1d;
    elseif dimension == 2
        truemodemap = [[
            i + (j - 1)*3 for i in truemodemap1d, j in truemodemap1d
        ]...];
    elseif dimension == 3
        truemodemap = [[
            i +
            (j - 1)*3 +
            (k - 1)*3^2
            for i in truemodemap1d, j in truemodemap1d, k in truemodemap1d
        ]...];
    end

    @assert truemodemap == modemap
end

# output

LFAToolkit.getnumberelements — Function

getnumberelements(basis)

Get the number of elements for the basis

Arguments:

basis: basis to compute number of micro-elements

Returns:

Integer number of basis micro-elements

Example:

# get number of nodes for basis
basis = TensorH1LagrangeMacroBasis(4, 4, 1, 2, 2);

# note: either syntax works
numbernodes = LFAToolkit.getnumberelements(basis);
numbernodes = basis.numberelements;

# verify
@assert numbernodes == 2^2

# output

LFAToolkit.getdiagonal — Function

getdiagonal(operator)

Compute or retrieve the symbol matrix diagonal for an operator

Arguments:

operator: operator to compute diagonal

Returns:

Symbol matrix diagonal for the operator

Example:

# setup
mesh = Mesh1D(1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);

# note: either syntax works
diagonal = LFAToolkit.getdiagonal(diffusion);
diagonal = diffusion.diagonal;

# verify
@assert diagonal ≈ [7/6 0; 0 4/3]

# output

LFAToolkit.getdXdxgradient — Function

getdXdxgradient(basis, mesh)

Get gradient adjusted for mesh stretching

Arguments:

basis: basis to compute gradient

mesh: mesh to compute gradient

Returns:

gradient matrix multiplied by change of coordinates adjoint

Example:

for dimension = 1:3
    # mesh
    mesh = []
    if dimension == 1
        mesh = Mesh1D(2.0)
    elseif dimension == 2
        mesh = Mesh2D(2.0, 3.0)
    elseif dimension == 3
        mesh = Mesh3D(2.0, 3.0, 4.0)
    end

    # basis
    basis = TensorH1LagrangeBasis(4, 3, 1, dimension);

    # get gradient on mesh
    gradient = LFAToolkit.getdXdxgradient(basis, mesh);

    # verify
    nodes = basis.nodes;
    linearfunction = [];
    truegradient = [];
    if dimension == 1
        linearfunction = nodes;
        truegradient = [1/2*ones(basis.numberquadraturepoints)...]
    elseif dimension == 2
        linearfunction = (nodes[:, 1] + nodes[:, 2]);
        truegradient = [
            1/2*ones(basis.numberquadraturepoints)...
            1/3*ones(basis.numberquadraturepoints)...
        ]
    elseif dimension == 3
        linearfunction = (nodes[:, 1] + nodes[:, 2] + nodes[:, 3]);
        truegradient = [
            1/2*ones(basis.numberquadraturepoints)...
            1/3*ones(basis.numberquadraturepoints)...
            1/4*ones(basis.numberquadraturepoints)...
        ]
    end

    @assert gradient*linearfunction ≈ truegradient
end

# output

LFAToolkit.getdxdXquadratureweights — Function

getdxdXquadratureweights(basis, mesh)

Get quadrature weights adjusted for mesh stretching

Arguments:

basis: basis to compute quadratureweights

mesh: mesh to compute quadratureweights

Returns:

quadrature weights multiplied by change of coordinates adjoint

Example:

mesh = Mesh1D(2.0)

# basis
basis = TensorH1LagrangeBasis(4, 3, 1, 1);

# get gradient on mesh
weights = LFAToolkit.getdxdXquadratureweights(basis, mesh);

# verify
@assert basis.quadratureweights * mesh.volume / basis.volume ≈ weights

# output

LFAToolkit.getprimalnodes — Method

getprimalnodes(basis)

Get primal nodes for basis

Arguments:

basis: basis to compute primal nodes

Returns:

Basis primal nodes vector

Example:

# test for all supported dimensions
for dimension in 1:3
    # get mode map vector
    p = 4
    basis = TensorH1LagrangeBasis(p, p, 1, dimension);

    # note: either syntax works
    primalnodes = LFAToolkit.getprimalnodes(basis);
    primalnodes = basis.primalnodes;

    # verify
    trueprimalnodes = []
    if dimension == 1
        trueprimalnodes = [1, p];
    elseif dimension == 2
        trueprimalnodes = [1, p, p^2-p+1, p^2]
    elseif dimension == 3
        trueprimalnodes = [1, p, p^2-p+1, p^2, p^3-p^2+1, p^3-p^2+p, p^3-p+1, p^3]
    end

    @assert trueprimalnodes == primalnodes
end

# output

LFAToolkit.getinterfacenodes — Method

getinterfacenodes(basis)

Get interface nodes for basis

Arguments:

basis: basis to compute primal nodes

Returns:

Basis primal nodes vector

Example:

# test for all supported dimensions
for dimension in 1:3
    # get mode map vector
    p = 4
    basis = TensorH1LagrangeBasis(p, p, 1, dimension);

    # note: either syntax works
    interfacenodes = LFAToolkit.getinterfacenodes(basis);
    interfacenodes = basis.interfacenodes;

    # verify
    trueinterfacenodes = []
    if dimension == 1
        trueinterfacenodes = [1, p];
    elseif dimension == 2
        trueinterfacenodes = [1:p..., p^2-p+1:p^2...]
        for i = 1:p-2
            push!(trueinterfacenodes, i*p+1)
            push!(trueinterfacenodes, (i+1)*p)
        end
    elseif dimension == 3
        trueinterfacenodes = [1:p^2..., p^3-p^2+1:p^3...]
        for i = 1:p-2
            push!(trueinterfacenodes, i*p^2+1:i*p^2+p...)
            push!(trueinterfacenodes, (i+1)*p^2-p+1:(i+1)*p^2...)
            push!(trueinterfacenodes, i*p^2+p+1:p:(i+1)*p^2-2*p+1...)
            push!(trueinterfacenodes, i*p^2+2*p:p:(i+1)*p^2-p...)
        end
    end
    trueinterfacenodes = sort(trueinterfacenodes)

    @assert trueinterfacenodes == interfacenodes
end

# output