Finite Element Operator

LFAToolkit.massoperator — Function

Operator("mass", p, mesh)

Convenience constructor for scalar mass operator

Arguments:

p1d: polynomial order of TensorH1LagrangeBasis

q1d: number of quadrature points in one dimension for basis

mesh: mesh for operator

Returns:

Mass matrix operator of order p on mesh

LFAToolkit.diffusionoperator — Function

Operator("diffusion", p, mesh)

Convenience constructor for scalar diffusion operator

Arguments:

p1d: polynomial order of TensorH1LagrangeBasis

q1d: number of quadrature points in one dimension for basis

mesh: mesh for operator

Returns:

Diffusion operator of order p on mesh

LFAToolkit.getelementmatrix — Function

getelementmatrix(operator)

Compute or retrieve the element matrix of operator for computing the symbol

Arguments:

operator: operator to compute element element matrix

Returns:

Assembled element matrix

Mass matrix example:

# setup
mesh = Mesh2D(1.0, 1.0);
mass = GalleryOperator("mass", 4, 4, mesh);
    
# element matrix computation
# note: either syntax works
elementmatrix = LFAToolkit.getelementmatrix(mass);
elementmatrix = mass.elementmatrix;

# verify
u = ones(4*4);
v = elementmatrix*u;
    
total = sum(v);
@assert total ≈ 1.0
    
# output

Diffusion matrix example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 4, 4, mesh);
    
# element matrix computation
# note: either syntax works
elementmatrix = LFAToolkit.getelementmatrix(diffusion);
elementmatrix = diffusion.elementmatrix;
    
# verify
u = ones(4*4);
v = elementmatrix*u;
    
total = sum(v);
@assert abs(total) < 1e-14
    
# output

LFAToolkit.getmultiplicity — Function

getmultiplicity(operator)

Compute or retrieve the vector of node multiplicity for the operator

Returns:

Vector of node multiplicity for the operator

Example

```jldoctest for dimension in 1:3 # setup mesh = [] if dimension == 1 mesh = Mesh1D(1.0); elseif dimension == 2 mesh = Mesh2D(1.0, 1.0); elseif dimension == 3 mesh = Mesh3D(1.0, 1.0, 1.0); end diffusion = GalleryOperator("diffusion", 5, 5, mesh);

# compute multiplicity
mult = diffusion.multiplicity;

# verify
mult1D = [2. 1. 1. 1. 2.];
if dimension == 1
    @assert mult ≈ mult1D
elseif dimension == 2
    @assert mult ≈ kron(mult1D, mult1D)
elseif dimension == 3
    @assert mult ≈ kron(mult1D, mult1D, mult1D)
end

end

LFAToolkit.getrowmodemap — Function

getrowmodemap(operator)

Compute or retrieve the matrix mapping the rows of the element matrix to the symbol matrix

Returns:

Matrix mapping rows of element matrix to symbol matrix

Example:

# setup
mesh = Mesh1D(1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# note: either syntax works
modemap = LFAToolkit.getrowmodemap(mass);
modemap = mass.rowmodemap;

# verify
@assert modemap ≈ [1 0 0 1; 0 1 0 0; 0 0 1 0]
    
# output

LFAToolkit.getcolumnmodemap — Function

getcolumnmodemap(operator)

Compute or retrieve the matrix mapping the columns of the element matrix to the symbol matrix

Returns:

Matrix mapping columns of element matrix to symbol matrix

Example:

# setup
mesh = Mesh1D(1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# note: either syntax works
modemap = LFAToolkit.getcolumnmodemap(mass);
modemap = mass.columnmodemap;

# verify
@assert modemap ≈ [1 0 0; 0 1 0; 0 0 1; 1 0 0]
    
# output

LFAToolkit.getinputcoordinates — Function

getinputcoordinates(operator)

Compute or retrieve the array of input coordinates

Returns:

Array of input coordinates

LFAToolkit.getoutputcoordinates — Function

getoutputcoordinates(operator)

Compute or retrieve the array of output coordinates

Returns:

Array of output coordinates

LFAToolkit.getnodecoordinatedifferences — Method

getnodecoordinatedifferences(operator)

Compute or retrieve the array of differences in coordinates between nodes

Returns:

Array of differences in coordinates between nodes

Example:

# setup
mesh = Mesh1D(1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# note: either syntax works
nodedifferences = LFAToolkit.getnodecoordinatedifferences(mass);
nodedifferences = mass.nodecoordinatedifferences;

# verify
truenodes = LFAToolkit.lobattoquadrature(4, false);
truenodedifferences = [
    (truenodes[j] - truenodes[i])/2.0 for i in 1:4, j in 1:4
];
@assert nodedifferences ≈ truenodedifferences
 
# output