Preconditioner: BDDC

LFAToolkit.BDDC — Type

BDDC(operator, coarseoperator, smoother, prolongation)

BDDC preconditioner for finite element operators

Arguments:

operator: finite element operator to precondition

injectiontype: type of injection into subassembled space to use

Returns:

BDDC preconditioner object

LFAToolkit.getprimalnodes — Method

getprimalnodes(bddc)

Compute or retrieve the primal nodes for a BDDC preconditioner

Returns:

Vector of primal nodes for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
primalnodes = LFAToolkit.getprimalnodes(bddc);
primalnodes = bddc.primalnodes;

# verify
@assert primalnodes == [1, p, p^2 - p + 1, p^2]

# output

LFAToolkit.getsubassemblednodes — Method

getsubassemblednodes(bddc)

Compute or retrieve the subassembled nodes for a BDDC preconditioner

Returns:

Vector of subassembled nodes for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
subassemblednodes = LFAToolkit.getsubassemblednodes(bddc);
subassemblednodes = bddc.subassemblednodes;

# verify
@assert subassemblednodes == setdiff(1:p^2, [1, p, p^2-p+1, p^2])

# output

LFAToolkit.getinterfacenodes — Method

getinterfacenodes(bddc)

Compute or retrieve the interface nodes for a BDDC preconditioner

Returns:

Vector of interface nodes for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
interfacenodes = LFAToolkit.getinterfacenodes(bddc);
interfacenodes = bddc.interfacenodes;

# verify
trueinterfacenodes = [1:p..., p^2-p+1:p^2...]
for i = 1:p-2
    push!(trueinterfacenodes, i*p+1)
    push!(trueinterfacenodes, (i+1)*p)
end
trueinterfacenodes = sort(trueinterfacenodes)
@assert interfacenodes == trueinterfacenodes

# output

LFAToolkit.getinteriornodes — Method

getinteriornodes(bddc)

Compute or retrieve the interior nodes for a BDDC preconditioner

Returns:

Vector of interior nodes for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
interiornodes = LFAToolkit.getinteriornodes(bddc);
interiornodes = bddc.interiornodes;

# verify
trueinterfacenodes = [1:p..., p^2-p+1:p^2...]
for i = 1:p-2
    push!(trueinterfacenodes, i*p+1)
    push!(trueinterfacenodes, (i+1)*p)
end
trueinterfacenodes = sort(trueinterfacenodes)
@assert interiornodes == setdiff(1:p^2, trueinterfacenodes)

# output

LFAToolkit.getprimalmodes — Function

getprimalmodes(bddc)

Compute or retrieve the primal modes for a BDDC preconditioner

Returns:

Vector of primal modes for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
primalmodes = LFAToolkit.getprimalmodes(bddc);
primalmodes = bddc.primalmodes;

# verify
@assert primalmodes == [1]

# output

LFAToolkit.getsubassembledmodes — Function

getsubassembledmodes(bddc)

Compute or retrieve the subassembled modes for a BDDC preconditioner

Returns:

Vector of subassembled modes for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
subassembledmodes = LFAToolkit.getsubassembledmodes(bddc);
subassembledmodes = bddc.subassembledmodes;

# verify
@assert subassembledmodes == setdiff(1:(p-1)^2, [1])

# output

LFAToolkit.getinterfacemodes — Function

getinterfacemodes(bddc)

Compute or retrieve the interface modes for a BDDC preconditioner

Returns:

Vector of interface modes for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
interfacemodes = LFAToolkit.getinterfacemodes(bddc);
interfacemodes = bddc.interfacemodes;

# verify
trueinterfacemodes = [1:p-1...]
for i = 1:p-2
    push!(trueinterfacemodes, i*(p-1)+1)
end
trueinterfacemodes = sort(trueinterfacemodes)
@assert interfacemodes == trueinterfacemodes

# output

LFAToolkit.getinteriormodes — Function

getinteriormodes(bddc)

Compute or retrieve the interior modes for a BDDC preconditioner

Returns:

Vector of interior modes for BDDC preconditioner

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
interiormodes = LFAToolkit.getinteriormodes(bddc);
interiormodes = bddc.interiormodes;

# verify
@assert interiormodes == setdiff(1:(p-1)^2, bddc.interfacemodes)

# output

LFAToolkit.getsubassembledinverse — Function

getsubassembledinverse(bddc)

Compute or retrieve the solver for subdomain for a BDDC preconditioner

Returns:

Solver for subdomain

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
subassembledinverse = LFAToolkit.getsubassembledinverse(bddc);
subassembledinverse = bddc.subassembledinverse;

# verify
@assert subassembledinverse ≈ Matrix(diffusion.elementmatrix[bddc.subassemblednodes, bddc.subassemblednodes])^-1

# output

LFAToolkit.getinteriorinverse — Function

getinteriorinverse(bddc)

Compute or retrieve the solver for subdomain interior for a BDDC preconditioner

Returns:

Solver for subdomain interior

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
interiorinverse = LFAToolkit.getinteriorinverse(bddc);
interiorinverse = bddc.interiorinverse;

# verify
@assert interiorinverse ≈ Matrix(diffusion.elementmatrix[bddc.interiornodes, bddc.interiornodes])^-1

# output

LFAToolkit.getschur — Function

getschur(bddc)

Compute or retrieve the Schur complement matrix for a BDDC preconditioner

Returns:

Matrix for Schur complement

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
p = 4
diffusion = GalleryOperator("diffusion", p, p, mesh);
bddc = LumpedBDDC(diffusion)

# note: either syntax works
schur = LFAToolkit.getschur(bddc);
schur = bddc.schur;

# verify
@assert schur ≈ diffusion.elementmatrix[bddc.primalnodes, bddc.primalnodes] -
    diffusion.elementmatrix[bddc.primalnodes, bddc.subassemblednodes] *
    Matrix(diffusion.elementmatrix[bddc.subassemblednodes, bddc.subassemblednodes])^-1 *
    diffusion.elementmatrix[bddc.subassemblednodes, bddc.primalnodes]

# output

LFAToolkit.getprimalrowmodemap — Function

getprimalrowmodemap(bddc)

Compute or retrieve the matrix mapping the rows of the primal BDDC matrix to the primal symbol matrix

Returns:

Matrix mapping rows of primal BDDC matrix to primal symbol matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);
bddc = LumpedBDDC(diffusion);

# note: either syntax works
primalmodemap = LFAToolkit.getprimalrowmodemap(bddc);
primalmodemap = bddc.primalrowmodemap;

# verify
@assert primalmodemap ≈ [1 1 1 1]

# output

LFAToolkit.getprimalcolumnmodemap — Function

getprimalcolumnmodemap(bddc)

Compute or retrieve the matrix mapping the columns of the primal BDDC matrix to the primal symbol matrix

Returns:

Matrix mapping columns of primal BDDC matrix to primal symbol matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);
bddc = LumpedBDDC(diffusion);

# note: either syntax works
primalmodemap = LFAToolkit.getprimalcolumnmodemap(bddc);
primalmodemap = bddc.primalcolumnmodemap;

# verify
@assert primalmodemap ≈ [1; 1; 1; 1]

# output

LFAToolkit.getsubassembledrowmodemap — Function

getsubassembledrowmodemap(bddc)

Compute or retrieve the matrix mapping the rows of the subassembled BDDC matrix to the subassembled symbol matrix

Returns:

Matrix mapping rows of subassembled BDDC matrix to subassembled symbol matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);
bddc = LumpedBDDC(diffusion);

# note: either syntax works
subassembledmodemap = LFAToolkit.getsubassembledrowmodemap(bddc);
subassembledmodemap = bddc.subassembledrowmodemap;

# verify
@assert subassembledmodemap ≈ bddc.operator.rowmodemap[bddc.subassembledmodes, bddc.subassemblednodes]

# output

LFAToolkit.getsubassembledcolumnmodemap — Function

getsubassembledcolumnmodemap(bddc)

Compute or retrieve the matrix mapping the columns of the subassembled BDDC matrix to the subassembled symbol matrix

Returns:

Matrix mapping columns of subassembled BDDC matrix to subassembled symbol matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);
bddc = LumpedBDDC(diffusion);

# note: either syntax works
subassembledmodemap = LFAToolkit.getsubassembledcolumnmodemap(bddc);
subassembledmodemap = bddc.subassembledcolumnmodemap;

# verify
@assert subassembledmodemap ≈ bddc.operator.columnmodemap[bddc.subassemblednodes, bddc.subassembledmodes]

# output

LFAToolkit.getinterfacerowmodemap — Function

getinterfacerowmodemap(bddc)

Compute or retrieve the matrix mapping the rows of the interface BDDC matrix to the interface symbol matrix

Returns:

Matrix mapping rows of interface BDDC matrix to interface symbol matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);
bddc = LumpedBDDC(diffusion);

# note: either syntax works
interfacemodemap = LFAToolkit.getinterfacerowmodemap(bddc);
interfacemodemap = bddc.interfacerowmodemap;

# verify
@assert interfacemodemap ≈ bddc.operator.rowmodemap[bddc.interfacemodes, bddc.interfacenodes]

# output

LFAToolkit.getinterfacecolumnmodemap — Function

getinterfacecolumnmodemap(bddc)

Compute or retrieve the matrix mapping the columns of the interface BDDC matrix to the interface symbol matrix

Returns:

Matrix mapping columns of interface BDDC matrix to interface symbol matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);
bddc = LumpedBDDC(diffusion);

# note: either syntax works
interfacemodemap = LFAToolkit.getinterfacecolumnmodemap(bddc);
interfacemodemap = bddc.interfacecolumnmodemap;

# verify
@assert interfacemodemap ≈ bddc.operator.columnmodemap[bddc.interfacenodes, bddc.interfacemodes]

# output

LFAToolkit.getinteriorrowmodemap — Function

getinteriorrowmodemap(bddc)

Compute or retrieve the matrix mapping the rows of the interior BDDC matrix to the interior symbol matrix

Returns:

Matrix mapping rows of interior BDDC matrix to interior symbol matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);
bddc = LumpedBDDC(diffusion);

# note: either syntax works
interiormodemap = LFAToolkit.getinteriorrowmodemap(bddc);
interiormodemap = bddc.interiorrowmodemap;

# verify
@assert interiormodemap ≈ bddc.operator.rowmodemap[bddc.interiormodes, bddc.interiornodes]

# output

LFAToolkit.getinteriorcolumnmodemap — Function

getinteriorcolumnmodemap(bddc)

Compute or retrieve the matrix mapping the columns of the interior BDDC matrix to the interior symbol matrix

Returns:

Matrix mapping columns of interior BDDC matrix to interior symbol matrix

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 3, mesh);
bddc = LumpedBDDC(diffusion);

# note: either syntax works
interiormodemap = LFAToolkit.getinteriorcolumnmodemap(bddc);
interiormodemap = bddc.interiorcolumnmodemap;

# verify
@assert interiormodemap ≈ bddc.operator.columnmodemap[bddc.interiornodes, bddc.interiormodes]

# output

LFAToolkit.computesymbolsrestriction — Method

computesymbolsrestriction(bddc)

Compute or retrieve the restriction operator for the BDDC symbol matrix

Returns:

Matrix providing the restriction operator of BDDC symbol matrix

Example:

LFAToolkit.computesymbolsinjection — Method

computesymbolsinjection(bddc)

Compute or retrieve the injection operator for the BDDC symbol matrix

Returns:

Matrix providing the injection operator of BDDC symbol matrix

Example: