Preconditioner: Multigrid

LFAToolkit.Multigrid — Type

Multigrid(fineoperator, coarseoperator, smoother, prolongation, multigridtype)

Multigrid preconditioner for finite element operators

Arguments:

fineoperator: finite element operator to precondition

coarseoperator: coarse grid representation of finite element operator to precondition

smoother: error relaxation operator, such as Jacobi

prolongationbases: element prolongation bases from coarse to fine grid

Returns:

Multigrid preconditioner object

LFAToolkit.getnodecoordinatedifferences — Method

getnodecoordinateddifferences(multigrid)

Compute or retrieve the array of differences in coordinates between nodes

Returns:

Array of differences in coordinates between nodes

LFAToolkit.getprolongationmatrix — Function

getprolongationmatrix(multigrid)

Compute or retrieve the prolongation matrix

Returns:

Matrix prolonging from coarse grid to fine grid

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
ctofbasis = TensorH1LagrangeBasis(2, 3, 1, 2, lagrangequadrature=true);

# operators
finediffusion = GalleryOperator("diffusion", 3, 3, mesh);
coarsediffusion = GalleryOperator("diffusion", 2, 3, mesh);

# smoother
jacobi = Jacobi(finediffusion);

# preconditioner
multigrid = PMultigrid(finediffusion, coarsediffusion, jacobi, [ctofbasis]);

# verify
u = ones(ctofbasis.numbernodes);
v = multigrid.prolongationmatrix * u;
@assert v' ≈ [4. 2. 4. 2. 1. 2. 4. 2. 4.].^-1

# output

LFAToolkit.computesymbolsprolongation — Method

computesymbolsprolongation(multigrid, θ)

Compute the symbol matrix for a multigrid prolongation operator

Arguments:

multigrid: Multigrid operator to compute prolongation symbol matrix for

θ: Fourier mode frequency array (one frequency per dimension)

Returns:

Symbol matrix for the multigrid prolongation operator