Public API

This page documents the public API of the LFAToolkit.

Mesh

LFAToolkit.Mesh1D — Type

Mesh1D(dx)

One dimensional regular background mesh

Arguments:

dx: deformation in x dimension

Returns:

One dimensional mesh object

Example:

# generate 1D mesh
mesh = Mesh1D(1.0);

# verify
@assert mesh.dimension == 1
@assert mesh.dx ≈ 1.0

# output

LFAToolkit.Mesh2D — Type

Mesh2D(dx, dy)

Two dimensional regular background mesh

Arguments:

dx: deformation in x dimension

dy: deformation in y dimension

Returns:

Two dimensional mesh object

Example:

# generate 2D mesh
mesh = Mesh2D(1.0, 0.5);

# verify
@assert mesh.dimension == 2
@assert mesh.dx ≈ 1.0
@assert mesh.dy ≈ 0.5

# output

LFAToolkit.Mesh3D — Type

Mesh3D(dx, dy, dz)

Three dimensional regular background mesh

Arguments:

dx: deformation in x dimension

dy: deformation in y dimension

dz: deformation in z dimension

Returns:

Three dimensional mesh object

Example:

# generate 3D mesh
mesh = Mesh3D(1.0, 0.5, 0.3);

# verify
@assert mesh.dimension == 3
@assert mesh.dx ≈ 1.0
@assert mesh.dy ≈ 0.5
@assert mesh.dz ≈ 0.3

# output

Basis

LFAToolkit.TensorBasis — Type

TensorBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    dimension,
    nodes1d,
    quadraturepoints1d,
    quadratureweights1d,
    interpolation1d,
    gradient1d
)

Tensor product basis

Arguments:

numbernodes1d: number of nodes in 1 dimension

numberquadraturepoints1d: number of quadrature points in 1 dimension

dimension: dimension of the basis

nodes1d: coordinates of the nodes in 1 dimension

quadraturepoints1d: coordinates of the quadrature points in 1 dimension

quadratureweights1d: quadrature weights in 1 dimension

interpolation1d: interpolation matrix from nodes to quadrature points in 1 dimension

gradient1d: gradient matrix from nodes to quadrature points in 1 dimension

Returns:

Tensor product basis object

LFAToolkit.NonTensorBasis — Type

NonTensorBasis(
    numbernodes,
    numberquadraturepoints,
    dimension,
    nodes,
    quadraturepoints,
    quadratureweights,
    interpolation,
    gradient
)

Non-tensor basis

Arguments:

numbernodes: number of nodes

numberquadraturepoints: number of quadrature points

dimension: dimension of the basis

nodes: coordinates of the nodes

quadraturepoints: coordinates of the quadrature points

quadratureweights: quadrature weights

interpolation: interpolation matrix from nodes to quadrature points

gradient: gradient matrix from nodes to quadrature points

Returns:

Non-tensor product basis object

LFAToolkit.TensorH1LagrangeBasis — Function

TensorH1LagrangeBasis(
    numbernodes,
    numberquadraturepoints1d,
    dimension
)

Tensor product basis on Gauss-Lobatto points with Gauss-Legendre quadrature

Arguments:

numbernodes1d: number of Gauss-Lobatto nodes

numberquadraturepoints1d: number of Gauss-Legendre quadrature points

dimension: dimension of basis

Returns:

H1 Lagrange tensor product basis object

Example:

# generate H1 Lagrange tensor product basis
basis = TensorH1LagrangeBasis(4, 3, 2);

# verify
@assert basis.numbernodes1d == 4
@assert basis.numberquadratuepoints1d == 3
@assert basis.dimension == 2

# output

Basis Evaluation Mode

LFAToolkit.EvaluationMode.EvalMode — Type

Basis evaluation mode for operator inputs and outputs

Modes:

interpolation: values interpolated to quadrature points

gradient: derivatives evaluated at quadrature points

quadratureweights: quadrature weights

Example:

EvaluationMode.EvalMode

# output
Enum LFAToolkit.EvaluationMode.EvalMode:
interpolation = 0
gradient = 1
quadratureweights = 2

Operator Field

LFAToolkit.OperatorField — Type

OperatorField(
    basis,
    evaluationmodes
)

Finite Element operator input or output, with a basis and evaluation mode

Arguments:

basis: finite element basis for the field

evaluationmodes: array of basis evaluation modes, note that quadrature weights must be listed in a separate operator field

Returns:

Finite element operator field object

Example:

# basis
basis = TensorH1LagrangeBasis(4, 3, 2);

# quadrature weights field
weightsfield = OperatorField(basis, [EvaluationMode.quadratureweights]);

# verify
@assert weightsfield.basis == basis
@assert weightsfield.evaluationmodes[1] == EvaluationMode.quadratureweights

# input or output field
inputfield = OperatorField(basis, [
    EvaluationMode.interpolation,
    EvaluationMode.gradient,
]);

# verify
@assert inputfield.basis == basis
@assert inputfield.evaluationmodes[1] == EvaluationMode.interpolation
@assert inputfield.evaluationmodes[2] == EvaluationMode.gradient

# output

Operator

LFAToolkit.Operator — Type

Operator(
    weakform,
    mesh,
    inputs,
    outputs
)

Finite element operator comprising of a weak form and bases

Arguments:

weakform: user provided function that represents weak form at quadrature points

mesh: mesh object with deformation in each dimension

inputs: array of operator input fields

outputs: array of operator output fields

Returns:

Finite element operator object

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(4, 4, 2);
    
function massweakform(u::Array{Float64}, w::Array{Float64})
    v = u * w[1]
    return [v]
end
    
# mass operator
inputs = [
    OperatorField(basis, [EvaluationMode.interpolation]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.interpolation])];
mass = Operator(massweakform, mesh, inputs, outputs);

# verify
@assert mass.weakform == massweakform
@assert mass.mesh == mesh
    
# output