Public API

This page documents the public API of the LFAToolkit.

Mesh

LFAToolkit.Mesh1D — Type

Mesh1D(dx)

One dimensional regular background mesh

Arguments:

dx: deformation in x dimension

Returns:

One dimensional mesh object

Example:

# generate 1D mesh
mesh = Mesh1D(1.0);

# verify
println(mesh)

# output
1d mesh:
    dx: 1.0

LFAToolkit.Mesh2D — Type

Mesh2D(dx, dy)

Two dimensional regular background mesh

Arguments:

dx: deformation in x dimension

dy: deformation in y dimension

Returns:

Two dimensional mesh object

Example:

# generate 2D mesh
mesh = Mesh2D(1.0, 0.5);

# verify
println(mesh)

# output
2d mesh:
    dx: 1.0
    dy: 0.5

LFAToolkit.Mesh3D — Type

Mesh3D(dx, dy, dz)

Three dimensional regular background mesh

Arguments:

dx: deformation in x dimension

dy: deformation in y dimension

dz: deformation in z dimension

Returns:

Three dimensional mesh object

Example:

# generate 3D mesh
mesh = Mesh3D(1.0, 0.5, 0.3);

# verify
println(mesh)

# output
3d mesh:
    dx: 1.0
    dy: 0.5
    dz: 0.3

Basis

LFAToolkit.TensorBasis — Type

TensorBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    dimension,
    nodes1d,
    quadraturepoints1d,
    quadratureweights1d,
    interpolation1d,
    gradient1d
)

Tensor product basis

Arguments:

numbernodes1d: number of nodes in 1 dimension

numberquadraturepoints1d: number of quadrature points in 1 dimension

dimension: dimension of the basis

nodes1d: coordinates of the nodes in 1 dimension

quadraturepoints1d: coordinates of the quadrature points in 1 dimension

quadratureweights1d: quadrature weights in 1 dimension

interpolation1d: interpolation matrix from nodes to quadrature points in 1 dimension

gradient1d: gradient matrix from nodes to quadrature points in 1 dimension

Returns:

Tensor product basis object

LFAToolkit.NonTensorBasis — Type

NonTensorBasis(
    numbernodes,
    numberquadraturepoints,
    dimension,
    nodes,
    quadraturepoints,
    quadratureweights,
    interpolation,
    gradient
)

Non-tensor basis

Arguments:

numbernodes: number of nodes

numberquadraturepoints: number of quadrature points

dimension: dimension of the basis

nodes: coordinates of the nodes

quadraturepoints: coordinates of the quadrature points

quadratureweights: quadrature weights

interpolation: interpolation matrix from nodes to quadrature points

gradient: gradient matrix from nodes to quadrature points

Returns:

Non-tensor product basis object

LFAToolkit.TensorH1LagrangeBasis — Function

TensorH1LagrangeBasis(
    numbernodes,
    numberquadraturepoints1d,
    dimension
)

Tensor product basis on Gauss-Lobatto points with Gauss-Legendre quadrature

Arguments:

numbernodes1d: number of Gauss-Lobatto nodes

numberquadraturepoints1d: number of Gauss-Legendre quadrature points

dimension: dimension of basis

Returns:

H1 Lagrange tensor product basis object

Example:

# generate H1 Lagrange tensor product basis
basis = TensorH1LagrangeBasis(4, 3, 2);

# verify
println(basis)

# output
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 3
    dimension: 2

Basis Evaluation Mode

LFAToolkit.EvaluationMode.EvalMode — Type

Basis evaluation mode for operator inputs and outputs

Modes:

interpolation: values interpolated to quadrature points

gradient: derivatives evaluated at quadrature points

quadratureweights: quadrature weights

Example:

EvaluationMode.EvalMode

# output
Enum LFAToolkit.EvaluationMode.EvalMode:
interpolation = 0
gradient = 1
quadratureweights = 2

Operator Field

LFAToolkit.OperatorField — Type

OperatorField(
    basis,
    evaluationmodes
)

Finite Element operator input or output, with a basis and evaluation mode

Arguments:

basis: finite element basis for the field

evaluationmodes: array of basis evaluation modes, note that quadrature weights must be listed in a separate operator field

Returns:

Finite element operator field object

Example:

# basis
basis = TensorH1LagrangeBasis(4, 3, 2);

# quadrature weights field
weightsfield = OperatorField(basis, [EvaluationMode.quadratureweights]);

# verify
println(weightsfield)

# input or output field
inputfield = OperatorField(basis, [
    EvaluationMode.interpolation,
    EvaluationMode.gradient,
]);

# verify
println(inputfield)

# output
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 3
    dimension: 2
  evaluation mode:
    quadratureweights
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 3
    dimension: 2
  evaluation modes:
    interpolation
    gradient

Operator

LFAToolkit.Operator — Type

Operator(
    weakform,
    mesh,
    inputs,
    outputs
)

Finite element operator comprising of a weak form and bases

Arguments:

weakform: user provided function that represents weak form at quadrature points

mesh: mesh object with deformation in each dimension

inputs: array of operator input fields

outputs: array of operator output fields

Returns:

Finite element operator object

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(4, 4, 2);
    
function massweakform(u::Array{Float64}, w::Array{Float64})
    v = u * w[1]
    return [v]
end
    
# mass operator
inputs = [
    OperatorField(basis, [EvaluationMode.interpolation]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.interpolation])];
mass = Operator(massweakform, mesh, inputs, outputs);

# verify
println(mass)

# output
finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 4
    dimension: 2
  evaluation mode:
    interpolation
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 4
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 4
    dimension: 2
  evaluation mode:
    interpolation

LFAToolkit.computesymbols — Method

computesymbols(operator, θ_x, ...)

Compute or retrieve the symbol matrix for an operator

Returns:

Symbol matrix for the operator

1D Example:

# setup
mesh = Mesh1D(1.0);
basis = TensorH1LagrangeBasis(3, 4, 1);
    
function diffusionweakform(du::Array{Float64}, w::Array{Float64})
    dv = du * w[1]
    return [dv]
end
    
# mass operator
inputs = [
    OperatorField(basis, [EvaluationMode.gradient]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.gradient])];
diffusion = Operator(diffusionweakform, mesh, inputs, outputs);

# compute symbols
A = computesymbols(diffusion, π);

# verify
using LinearAlgebra;
eigenvalues = eigvals(A);
@assert eigenvalues ≈ [1; 4/3]
 
# output

2D Example:

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(3, 4, 2);
    
function diffusionweakform(du::Array{Float64}, w::Array{Float64})
    dv = du * w[1]
    return [dv]
end
    
# mass operator
inputs = [
    OperatorField(basis, [EvaluationMode.gradient]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.gradient])];
diffusion = Operator(diffusionweakform, mesh, inputs, outputs);

# compute symbols
A = computesymbols(diffusion, π, π);

# verify
using LinearAlgebra;
eigenvalues = eigvals(A);
@assert eigenvalues ≈ [1/3; 4/9; 16/30; 64/90]
 
# output

Preconditioners

Jacobi

LFAToolkit.Jacobi — Type

Jacobi(operator)

Jacobi diagonal preconditioner for finite element operators

Arguments:

operator: finite element operator to precondition

Returns:

Jacobi preconditioner object

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(4, 4, 2);
    
function massweakform(u::Array{Float64}, w::Array{Float64})
    v = u * w[1]
    return [v]
end
    
# mass operator
inputs = [
    OperatorField(basis, [EvaluationMode.interpolation]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.interpolation])];
mass = Operator(massweakform, mesh, inputs, outputs);

# preconditioner
jacobi = Jacobi(mass);

# verify
println(jacobi)
println(jacobi.operator)

# output
jacobi preconditioner
finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 4
    dimension: 2
  evaluation mode:
    interpolation
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 4
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 4
    dimension: 2
  evaluation mode:
    interpolation

LFAToolkit.computesymbols — Method

computesymbols(preconditioner, θ_x, ...)

Compute or retrieve the symbol matrix for a Jacobi preconditioned operator

Returns:

Symbol matrix for the Jacobi preconditioned operator

1D Example:

# setup
mesh = Mesh1D(1.0);
basis = TensorH1LagrangeBasis(3, 4, 1);
    
function diffusionweakform(du::Array{Float64}, w::Array{Float64})
    dv = du * w[1]
    return [dv]
end
    
# mass operator
inputs = [
    OperatorField(basis, [EvaluationMode.gradient]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.gradient])];
diffusion = Operator(diffusionweakform, mesh, inputs, outputs);

# preconditioner
jacobi = Jacobi(diffusion);

# note: either syntax works
A = computesymbols(jacobi, π);

# verify
using LinearAlgebra;
eigenvalues = eigvals(A);
@assert max(real(eigenvalues)...) ≈ 1.0
 
# output

2D Example:

# setup
mesh = Mesh2D(1.0, 1.0);
basis = TensorH1LagrangeBasis(3, 4, 2);
    
function diffusionweakform(du::Array{Float64}, w::Array{Float64})
    dv = du * w[1]
    return [dv]
end
    
# mass operator
inputs = [
    OperatorField(basis, [EvaluationMode.gradient]),
    OperatorField(basis, [EvaluationMode.quadratureweights]),
];
outputs = [OperatorField(basis, [EvaluationMode.gradient])];
diffusion = Operator(diffusionweakform, mesh, inputs, outputs);

# preconditioner
jacobi = Jacobi(diffusion)

# compute symbols
A = computesymbols(jacobi, π, π);

# verify
using LinearAlgebra;
eigenvalues = eigvals(A);
@assert max(real(eigenvalues)...) ≈ 5/4
 
# output