Finite Element Basis

The finite element basis objects represent the discretization of the problem.

There are convenience constructors for H1 tensor product bases on uniformly spaced, Gauss-Lobatto, or Gauss-Legendre points with Gauss-Lobatto or Gauss-Legendre quadrature. Users can create additional finite element bases if the prerequisite information is provided.

Base Classes

These are the bases classes for finite element bases. The constructors for these base classes can be used to create user defined finite elements.

LFAToolkit.TensorBasis — Type

TensorBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    numbercomponents,
    dimension,
    nodes1d,
    quadraturepoints1d,
    quadratureweights1d,
    interpolation1d,
    gradient1d
)

Tensor product basis

Arguments:

numbernodes1d: number of nodes in 1 dimension

numberquadraturepoints1d: number of quadrature points in 1 dimension

numbercomponents: number of components

dimension: dimension of the basis

nodes1d: coordinates of the nodes in 1 dimension

quadraturepoints1d: coordinates of the quadrature points in 1 dimension

quadratureweights1d: quadrature weights in 1 dimension

interpolation1d: interpolation matrix from nodes to quadrature points in 1 dimension

gradient1d: gradient matrix from nodes to quadrature points in 1 dimension

Returns:

Tensor product basis object

LFAToolkit.NonTensorBasis — Type

NonTensorBasis(
    numbernodes,
    numberquadraturepoints,
    numbercomponents,
    dimension,
    nodes,
    quadraturepoints,
    quadratureweights,
    interpolation,
    gradient
)

Non-tensor basis

Arguments:

numbernodes: number of nodes

numberquadraturepoints: number of quadrature points

numbercomponents: number of components

dimension: dimension of the basis

nodes: coordinates of the nodes

quadraturepoints: coordinates of the quadrature points

quadratureweights: quadrature weights

interpolation: interpolation matrix from nodes to quadrature points

gradient: gradient matrix from nodes to quadrature points

Returns:

Non-tensor product basis object

Singe Element Bases

These bases represent common tensor product finite element bases used for continuous Galerkin methods.

LFAToolkit.TensorH1LagrangeBasis — Function

TensorH1LagrangeBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    numbercomponents,
    dimension,
    lagrangequadrature = false,
)

Tensor product basis on Gauss-Lobatto points with Gauss-Legendre quadrature

Arguments:

numbernodes1d: number of Gauss-Lobatto nodes

numberquadraturepoints1d: number of Gauss-Legendre quadrature points

numbercomponents: number of components

dimension: dimension of basis

lagrangequadrature=false: Gauss-Lagrange or Gauss-Lobatto quadrature points, default: Gauss-Lobatto

Returns:

H1 Lagrange tensor product basis object

Example:

# generate H1 Lagrange tensor product basis
basis = TensorH1LagrangeBasis(4, 4, 3, 2);

# generate basis with Lagrange quadrature points
basis = TensorH1LagrangeBasis(4, 4, 3, 2, lagrangequadrature=true);

# verify
println(basis)

# output
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 4
    numbercomponents: 3
    dimension: 2

LFAToolkit.TensorH1UniformBasis — Function

TensorH1UniformBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    numbercomponents,
    dimension,
)

Tensor product basis on uniformly points with Gauss-Legendre quadrature

Arguments:

numbernodes1d: number of uniformly spaced nodes

numberquadraturepoints1d: number of Gauss-Legendre quadrature points

numbercomponents: number of components

dimension: dimension of basis

Returns:

H1 uniformly spaced tensor product basis object

Example:

# generate H1 uniformly spaced tensor product basis
basis = TensorH1UniformBasis(4, 3, 2, 1);

# verify
println(basis)

# output
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 3
    numbercomponents: 2
    dimension: 1

Macro-Element Bases

These bases represent a macro-element consisting of multiple overlapping micro elements, where the micro elements are created as above.

LFAToolkit.TensorH1LagrangeMacroBasis — Function

TensorH1LagrangeMacroBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    numbercomponents,
    dimension,
    numberelements1d,
    lagrangequadrature,
)

Tensor product macro-element basis on Gauss-Lobatto points with Gauss-Legendre quadrature

Arguments:

numbernodes1d: number of Gauss-Lobatto nodes

numberquadraturepoints1d: number of Gauss-Legendre quadrature points

numbercomponents: number of components

dimension: dimension of basis

numberelements1d: number of elements in macro-element

lagrangequadrature=false: Gauss-Lagrange or Gauss-Lobatto quadrature points, default: Gauss-Lobatto

Returns:

H1 Lagrange tensor product macro-element basis object

Example:

# generate H1 Lagrange tensor macro-element product basis
basis = TensorH1LagrangeMacroBasis(4, 4, 1, 2, 2);

# generate basis with Lagrange quadrature points
basis = TensorH1LagrangeMacroBasis(4, 4, 1, 2, 2, lagrangequadrature=true);

# verify
println(basis)

# output
macro-element tensor product basis:
    numbernodes1d: 7
    numberquadraturepoints1d: 8
    numbercomponents: 1
    numberelements1d: 2
    dimension: 2

LFAToolkit.TensorH1UniformMacroBasis — Function

TensorH1UniformMacroBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    numbercomponents,
    dimension,
    numberelements1d,
)

Tensor product macro-element basis on uniformly points with Gauss-Legendre quadrature

Arguments:

numbernodes1d: number of uniformly spaced nodes

numberquadraturepoints1d: number of Gauss-Legendre quadrature points

numbercomponents: number of components

dimension: dimension of basis

numberelements1d: number of elements in macro-element

Returns:

H1 uniformly spaced tensor product macro-element basis object

Example:

# generate H1 uniformly spaced tensor product macro-element basis
basis = TensorH1UniformMacroBasis(4, 3, 1, 2, 2);

# verify
println(basis)

# output
macro-element tensor product basis:
    numbernodes1d: 7
    numberquadraturepoints1d: 6
    numbercomponents: 1
    numberelements1d: 2
    dimension: 2

P Prolongation Basis

This basis provides prolongation from a single element to a single element of higher order.

LFAToolkit.TensorH1LagrangePProlongationBasis — Function

TensorH1LagrangePProlongationBasis(
    numbercoarsenodes1d,
    numberfinenodes1d,
    numbercomponents,
    dimension,
)

Tensor product p prolongation basis on Gauss-Lobatto points

Arguments:

numbercoarsenodes1d: number of coarse grid Gauss-Lobatto nodes

numberfinenodes1d: number of fine grid Gauss-Lobatto nodes

numbercomponents: number of components

dimension: dimension of basis

Returns:

H1 Lagrange tensor product basis object

Example:

# generate H1 Lagrange tensor product basis
basisctof = TensorH1LagrangePProlongationBasis(2, 3, 1, 2);

# verify
println(basisctof)

# output
tensor product basis:
    numbernodes1d: 2
    numberquadraturepoints1d: 3
    numbercomponents: 1
    dimension: 2

H Prolongation Bases

These bases provide prolongation from a single element to a macro-element of the same order or a macro-element to a macro-element with a larger number of elements.

LFAToolkit.TensorH1LagrangeHProlongationBasis — Function

TensorH1LagrangeHProlongationBasis(
    numbernodes1d,
    numbercomponents,
    dimension,
    numberfineelements1d,
)

Tensor product h prolongation basis on Gauss-Lobatto points

Arguments:

numbernodes1d: number of Gauss-Lobatto nodes per element

numbercomponents: number of components

dimension: dimension of basis

numberfineelements1d: number of fine grid elements

Returns:

H1 Gauss-Lobatto tensor product h prolongation basis object

Example:

# generate H1 Gauss-Lobatto tensor product h prolongation basis
basis = TensorH1LagrangeHProlongationBasis(4, 3, 2, 2);

# verify
println(basis)

# output
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 7
    numbercomponents: 3
    dimension: 2

LFAToolkit.TensorH1UniformHProlongationBasis — Function

TensorH1UniformHProlongationBasis(
    numbernodes1d,
    numbercomponents,
    dimension,
    numberfineelements1d,
)

Tensor product h prolongation basis on uniformly spaced points

Arguments:

numbernodes1d: number of uniformly spaced nodes per element

numbercomponents: number of components

dimension: dimension of basis

numberfineelements1d: number of fine grid elements

Returns:

H1 uniformly spaced tensor product h prolongation basis object

Example:

# generate H1 uniformly spaced tensor product h prolongation basis
basis = TensorH1UniformHProlongationBasis(4, 3, 2, 2);

# verify
println(basis)

# output
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 7
    numbercomponents: 3
    dimension: 2

LFAToolkit.TensorH1LagrangeHProlongationMacroBasis — Function

TensorH1LagrangeHProlongationMacroBasis(
    numbernodes1d,
    numbercomponents,
    dimension,
    numbercoarseelements1d,
    numberfineelements1d,
)

Tensor product macro-element h prolongation basis on Gauss-Lobatto points

Arguments:

numbernodes1d: number of Gauss-Lobatto nodes per element

numbercomponents: number of components

dimension: dimension of basis

numbercoarseelements1d: number of coarse grid elements in macro-element

numberfineelements1d: number of fine grid elements in macro-element

Returns:

H1 Gauss-Lobatto tensor product h prolongation macro-element basis object

Example:

# generate H1 Gauss-Lobatto tensor product h prolongation macro-element basis
basis = TensorH1LagrangeHProlongationMacroBasis(4, 1, 2, 2, 4);

# verify
println(basis)

# output
macro-element tensor product basis:
    numbernodes1d: 7
    numberquadraturepoints1d: 13
    numbercomponents: 1
    numberelements1d: 2
    dimension: 2

LFAToolkit.TensorH1UniformHProlongationMacroBasis — Function

TensorH1UniformHProlongationMacroBasis(
    numbernodes1d,
    numbercomponents,
    dimension,
    numbercoarseelements1d,
    numberfineelements1d,
)

Tensor product macro-element h prolongation basis on uniformly spaced points

Arguments:

numbernodes1d: number of uniformly spaced nodes per element

numbercomponents: number of components

dimension: dimension of basis

numbercoarseelements1d: number of coarse grid elements in macro-element

numberfineelements1d: number of fine grid elements in macro-element

Returns:

H1 uniformly spaced tensor product h prolongation macro-element basis object

Example:

# generate H1 uniformly spaced tensor product h prolongation macro-element basis
basis = TensorH1UniformHProlongationMacroBasis(4, 1, 2, 2, 4);

# verify
println(basis)

# output
macro-element tensor product basis:
    numbernodes1d: 7
    numberquadraturepoints1d: 13
    numbercomponents: 1
    numberelements1d: 2
    dimension: 2