Finite Element Basis

LFAToolkit.TensorBasis — Type

TensorBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    dimension,
    nodes1d,
    quadraturepoints1d,
    quadratureweights1d,
    interpolation1d,
    gradient1d
)

Tensor product basis

Arguments:

numbernodes1d: number of nodes in 1 dimension

numberquadraturepoints1d: number of quadrature points in 1 dimension

dimension: dimension of the basis

nodes1d: coordinates of the nodes in 1 dimension

quadraturepoints1d: coordinates of the quadrature points in 1 dimension

quadratureweights1d: quadrature weights in 1 dimension

interpolation1d: interpolation matrix from nodes to quadrature points in 1 dimension

gradient1d: gradient matrix from nodes to quadrature points in 1 dimension

Returns:

Tensor product basis object

LFAToolkit.NonTensorBasis — Type

NonTensorBasis(
    numbernodes,
    numberquadraturepoints,
    dimension,
    nodes,
    quadraturepoints,
    quadratureweights,
    interpolation,
    gradient
)

Non-tensor basis

Arguments:

numbernodes: number of nodes

numberquadraturepoints: number of quadrature points

dimension: dimension of the basis

nodes: coordinates of the nodes

quadraturepoints: coordinates of the quadrature points

quadratureweights: quadrature weights

interpolation: interpolation matrix from nodes to quadrature points

gradient: gradient matrix from nodes to quadrature points

Returns:

Non-tensor product basis object

LFAToolkit.TensorH1LagrangeBasis — Function

TensorH1LagrangeBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    dimension,
    collocatedquadrature,
)

Tensor product basis on Gauss-Lobatto points with Gauss-Legendre quadrature

Arguments:

numbernodes1d: number of Gauss-Lobatto nodes

numberquadraturepoints1d: number of Gauss-Legendre quadrature points

dimension: dimension of basis

Returns:

H1 Lagrange tensor product basis object

Example:

# generate H1 Lagrange tensor product basis
basis = TensorH1LagrangeBasis(4, 4, 2);

# generate basis with Lagrange quadrature points
lagrangequadrature = true;
basis = TensorH1LagrangeBasis(4, 4, 2, lagrangequadrature);

# verify
println(basis)

# output
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 4
    dimension: 2

LFAToolkit.TensorH1UniformBasis — Function

TensorH1UniformBasis(
    numbernodes1d,
    numberquadraturepoints1d,
    dimension,
)

Tensor product basis on uniformly points with Gauss-Legendre quadrature

Arguments:

numbernodes1d: number of uniformly spaced nodes

numberquadraturepoints1d: number of Gauss-Legendre quadrature points

dimension: dimension of basis

Returns:

H1 uniformly spaced tensor product basis object

Example:

# generate H1 uniformly spaced tensor product basis
basis = TensorH1UniformBasis(4, 3, 2);

# verify
println(basis)

# output
tensor product basis:
    numbernodes1d: 4
    numberquadraturepoints1d: 3
    dimension: 2