Finite Element Operator Gallery

The following PDEs are available in the operator gallery.

Scalar Mass Operator

LFAToolkit.massoperator — Function

GalleryOperator("mass", p1d, q1d, mesh)

Convenience constructor for scalar mass operator

Weak form:

$\int v u$

Arguments:

p1d: polynomial order of TensorH1LagrangeBasis

q1d: number of quadrature points in one dimension for basis

mesh: mesh for operator

Returns:

Mass matrix operator of order p on mesh

Example:

# mass operator
mesh = Mesh2D(1.0, 1.0);
mass = GalleryOperator("mass", 3, 4, mesh);

# verify
println(mass)

# output
finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    interpolation
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    quadratureweights
   
1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    interpolation

Vector Mass Operator

LFAToolkit.vectormassoperator — Function

GalleryOperator("vectormass", p1d, q1d, mesh)

Convenience constructor for vector mass operator in three components

Weak form:

$\int \mathbf{v} \mathbf{u}$

Arguments:

p1d: polynomial order of TensorH1LagrangeBasis

q1d: number of quadrature points in one dimension for basis

mesh: mesh for operator

Returns:

Vector mass matrix operator of order p on mesh

Example:

# mass operator
mesh = Mesh2D(1.0, 1.0);
mass = GalleryOperator("vectormass", 3, 4, mesh);

# verify
println(mass)

# output
finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 3
    dimension: 2
  evaluation mode:
    interpolation
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 3
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 3
    dimension: 2
  evaluation mode:
    interpolation

Scalar Diffusion Operator

LFAToolkit.diffusionoperator — Function

GalleryOperator("diffusion", p1d, q1d, mesh)

Convenience constructor for scalar diffusion operator

Weak form:

$\int \nabla v \nabla u$

Arguments:

p1d: polynomial order of TensorH1LagrangeBasis

q1d: number of quadrature points in one dimension for basis

mesh: mesh for operator

Returns:

Diffusion operator of order p on mesh

Example:

# mass operator
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("diffusion", 3, 4, mesh);

# verify
println(diffusion)

# output

finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    gradient
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    quadratureweights
   
1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 1
    dimension: 2
  evaluation mode:
    gradient

Vector Diffusion Operator

LFAToolkit.vectordiffusionoperator — Function

GalleryOperator("vectordiffusion", p1d, q1d, mesh)

Convenience constructor for vector diffusion operator in three components

Weak form:

$\int \nabla \mathbf{v} \nabla \mathbf{u}$

Arguments:

p1d: polynomial order of TensorH1LagrangeBasis

q1d: number of quadrature points in one dimension for basis

mesh: mesh for operator

Returns:

Vector diffusion operator of order p on mesh

Example:

# mass operator
mesh = Mesh2D(1.0, 1.0);
diffusion = GalleryOperator("vectordiffusion", 3, 4, mesh);

# verify
println(diffusion)

# output

finite element operator:
2d mesh:
    dx: 1.0
    dy: 1.0

2 inputs:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 3
    dimension: 2
  evaluation mode:
    gradient
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 3
    dimension: 2
  evaluation mode:
    quadratureweights

1 output:
operator field:
  tensor product basis:
    numbernodes1d: 3
    numberquadraturepoints1d: 4
    numbercomponents: 3
    dimension: 2
  evaluation mode:
    gradient