Preconditioner: Identity

The identity preconditioner object allows investigation of multigrid methods without smoothers.

Documentation

LFAToolkit.IdentityPC — Type

IdentityPC()

Identity preconditioner to investigate multigrid without smoother

Returns:

identity preconditioner object

Example:

# setup
mesh = Mesh2D(1.0, 1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# preconditioner
identity = IdentityPC(mass);

# verify
println(identity)

# output

identity preconditioner

LFAToolkit.computesymbols — Method

computesymbols(identity, ω, θ)

Compute or retrieve the symbol matrix for a identity preconditioned operator

Arguments:

identity::IdentityPC: Identity preconditioner to compute symbol matrix for

ω::Array: smoothing weighting factor array

θ::Array: Fourier mode frequency array (one frequency per dimension)

Returns:

symbol matrix for the identity preconditioner (I)

Example:

using LinearAlgebra

# setup
mesh = Mesh2D(1.0, 1.0);
mass = GalleryOperator("mass", 4, 4, mesh);

# preconditioner
identity = IdentityPC(mass);

# compute symbols
A = computesymbols(identity, [], []);

# verify
@assert A ≈ I

# output